% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_baseline_data.R
\name{add_baseline_data<-}
\alias{add_baseline_data<-}
\title{Adding baseline data to a rtmle object}
\usage{
add_baseline_data(x, ...) <- value
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{...}{Not used (not yet)}

\item{value}{data.frame, data.table or tibble which contains the baseline data}
}
\description{
This function adds a baseline dataset to an existing rtmle object
}
