% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheap_bootstrap.R
\name{cheap_bootstrap_rtmle}
\alias{cheap_bootstrap_rtmle}
\title{Perform the Cheap Subsampling Procedure for LTMLE}
\usage{
cheap_bootstrap_rtmle(
  x,
  cheap_bootstrap_arguments = list(b = 25, type = "subsampling", size = NULL),
  rtmle_arguments,
  id_name
)
}
\arguments{
\item{x}{An object contaiining a prepared \code{rtmle} object}

\item{cheap_bootstrap_arguments}{A list specifying bootstrap parameters given to the \code{cheap_bootstrap} function.
\itemize{
\item \code{b} (integer): Number of bootstrap samples (default: 25).
\item \code{type} (character): Type of resampling method, e.g., "subsampling".
\item \code{size} (integer, optional): Size of each subsample
}}

\item{rtmle_arguments}{A list of arguments passed to the \code{run_rtmle} function.}

\item{id_name}{A character string specifying the name of the identifier column used for personal ids.}
}
\description{
This function applies a bootstrap resampling technique using the \code{CheapSubsampling} package.
It modifies the provided dataset \code{x} by resampling observations and then applies LTMLE.
}
\examples{

\dontrun{
set.seed(17)
tau <- 3
ld <- simulate_long_data(n = 100,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(intervals = tau,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x$long_data <- ld[c("outcome_data","censored_data","competing_data","timevar_data")]
add_baseline_data(x) <- ld$baseline_data[,start_followup_date:=0]
x <- long_to_wide(x,intervals = seq(0,2000,30.45*12))
protocol(x) <- list(name = "Always_A",
                    intervention = data.frame("A" = factor("1",levels = c("0","1"))))
protocol(x) <- list(name = "Never_A",
                    intervention = data.frame("A" = factor("0",levels = c("0","1"))))
prepare_data(x) <- list()
target(x) <- list(name = "Outcome_risk",
                  strategy = "additive",
                  estimator = "tmle",
                  protocols = c("Always_A","Never_A"))
cb <- cheap_bootstrap_rtmle(
 x = x,
rtmle_arguments = list(learner = "learn_glm",time_horizon = 1:tau),
  id_name = "id"
)
x <- run_rtmle(x,learner = "learn_glm",time_horizon = 1:tau)
summary(x)
cb
}

}
