% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_glmnet.R
\name{learn_glmnet}
\alias{learn_glmnet}
\title{Nuisance-parameter learner based on glmnet}
\usage{
learn_glmnet(
  character_formula,
  data,
  intervened_data,
  selector = "undersmooth",
  lambda = NULL,
  alpha = 1,
  family,
  nfolds = 10,
  type.measure = "deviance",
  save_fitted_objects = FALSE,
  reuse_fit = NULL,
  ...
)
}
\arguments{
\item{character_formula}{Formula for the nuisance parameter, supplied as a
character string.}

\item{data}{Data used for learning.}

\item{intervened_data}{Data used for prediction after intervention variables
have been set according to a regime.}

\item{selector}{Character value controlling how the penalty parameter
\code{lambda} is selected. If \code{"undersmooth"}, use the least
penalized value that still fits the model. The other options,
\code{"min"} and \code{"1se"}, are described in
\code{\link[glmnet]{cv.glmnet}}.}

\item{lambda}{Penalty parameter passed to \code{\link[glmnet]{glmnet}} or
\code{\link[glmnet]{cv.glmnet}}.}

\item{alpha}{Elastic-net mixing parameter. See
\code{\link[glmnet]{glmnet}}.}

\item{family}{Passed to \code{\link[glmnet]{glmnet}} or
\code{\link[glmnet]{cv.glmnet}}.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{type.measure}{Loss function used for cross-validation. Default is
deviance.}

\item{save_fitted_objects}{Logical. If \code{TRUE}, store the 
fitted object as element \code{fit}}

\item{reuse_fit}{Optional fitted object returned by a previous call. If
supplied, skip fitting and only predict in \code{intervened_data}.}

\item{...}{Additional arguments passed to \code{\link[glmnet]{glmnet}}.}
}
\value{
A list whose first element, \code{predicted_values}, is a vector of
  predicted probabilities. Element \code{fit} contains the selected
  coefficients, and element \code{object} contains the fitted model.
}
\description{
Learn nuisance-parameter models with glmnet
}
\details{
Learns nuisance-parameter models for TMLE and predicts probabilities in
intervention-updated data using \code{\link[glmnet]{glmnet}}.
}
\examples{
d <- data.table::data.table(Y = rep(c(0, 1), 10),
                            A = rep(c(0, 1, 1, 0), 5),
                            L = seq(-1, 1, length.out = 20))
predicted <- learn_glmnet("Y ~ A + L", data = d, intervened_data = d,
                          lambda = 0.01)
head(predicted$predicted_values)
predicted$selected.lambda
}
\seealso{
\code{\link{superlearn}}, \code{\link{learn_ranger}},
  \code{\link{learn_glm}}, \code{\link{learn_xgboost}}
}
