% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_wide.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Obsolete alias for \code{\link{discretize}}}
\usage{
long_to_wide(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{\link{discretize}}.}
}
\value{
The modified \code{rtmle} object returned by \code{discretize}.
}
\description{
\code{long_to_wide()} has been renamed to \code{discretize()}.
This alias is retained for existing scripts and forwards all arguments to
the new function.
}
\seealso{
\code{\link{discretize}}, \code{\link{map_data_to_grid}}
}
