% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare}
\alias{plot_compare}
\title{Plot comparison of estimates.}
\usage{
plot_compare(estimates, plot_no = 1, diff_betas = seq(0, 1, by = 0.1))
}
\arguments{
\item{estimates}{A matrix B times 6 matrix with estimates data where A0 = 0 or 1
and where the effect of the drug varies from no effect, small effect to large effect.}

\item{plot_no}{For now the possibility is 1 or 2. Plot_no = 1 lets us plot the different
estimates. While plot_no = 2, lets us plot differences.}

\item{diff_betas}{The values of the varying beta coefficient.}
}
\value{
Plot
}
\description{
\code{plot_compare} is a function for plotting an estimates. It takes a matrix of estimates, and plots.
}
\examples{
# First we define an estimator
estimator1 <- function(data, N) {

T2D_events <- data[Delta == 3]
return(c(nrow(T2D_events[A0 == 1 & Time < 5])/N, nrow(T2D_events[A0 == 0 & Time < 5])/N))

}

res1 <- compare_effects(estimator = estimator1, N = 100, beta_L_D = seq(0,1,by = 0.1))
plot_compare(res1, diff_betas = seq(0,1,by = 0.1))
}
