% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recur_event_data.R
\name{plot_recur_event_data}
\alias{plot_recur_event_data}
\title{Plot Recurrent Event Data}
\usage{
plot_recur_event_data(data)
}
\arguments{
\item{data}{Survival data: a data frame containing an ID, Time, Delta and L0 column. Possibly a L1 and A column.}
}
\value{
A plot of the survival data
}
\description{
Plot Recurrent Event Data
}
\examples{
data <- sim_recur_event_data(10)
plot_recur_event_data(data)
}
