% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data<-}
\alias{prepare_data<-}
\title{Preparing a targeted minimum loss based analysis}
\usage{
prepare_data(x, ...) <- value
}
\arguments{
\item{x}{Object which contains long format or wide format data}

\item{...}{not used}

\item{value}{A list with the following forced elements:
\itemize{
\item \code{"intervals"} The time intervals
}}
}
\value{
The object augmented with a new element called \code{prepared_data}.
}
\description{
If the object contains data in long format these are
transformed to wide format. The wide format data are sorted and
checked for the analysis and added as \code{x$prepared_data}.
}
\details{
As a wanted side effect the function identifies the variables
for the intervention nodes (A_nodes), the time dependent covariates (B_nodes)
the outcome nodes (Y_nodes), the competing risk nodes (D_nodes) and the
censoring nodes (C_nodes) and adds them to the object.
}
\examples{
set.seed(112)
ld <- simulate_long_data(n = 101,number_visits = 20,
                         beta = list(A_on_Y = -.2,
                         A0_on_Y = -0.3,A0_on_A = 6),
                               register_format = TRUE)
x <- rtmle_init(intervals = 8, name_id = "id",
                name_outcome = "Y", name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x$long_data <- ld[c("outcome_data","censored_data","competing_data","timevar_data")]
add_baseline_data(x) <- ld$baseline_data[,start_followup_date:=0]
x <- long_to_wide(x,intervals=seq(0,2000,30.45*6))
prepare_data(x) <- list()
                       
x$prepared_data$data
}
\seealso{
rtmle_init
}
\author{
Thomas A. Gerds \href{mailto:tag@biostat.ku.dk}{tag@biostat.ku.dk}
}
