% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rtmle.R
\name{print.rtmle}
\alias{print.rtmle}
\title{Print an rtmle object}
\usage{
\method{print}{rtmle}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Not used.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Prints the current state of an \code{rtmle} object and indicates the next
analysis steps that are still missing.
}
\examples{
x <- rtmle_init(time_grid = 0:2,
                name_id = "id",
                name_outcome = "Y")
print(x)
}
\seealso{
\code{\link{rtmle_init}}, \code{\link{prepare_rtmle_data}},
  \code{\link{run_rtmle}}, \code{\link{summary.rtmle}}
}
