% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol<-}
\alias{protocol<-}
\title{Define a named protocol for a hypothetical/emulated trial}
\usage{
protocol(x, ...) <- value
}
\arguments{
\item{x}{An rtmle object as obtained with \code{rtmle_init}.}

\item{...}{Not (yet) used}

\item{value}{list with three forced elements:
\itemize{
\item \code{name}: the name of the protocol
\item \code{treatment_variable}: A list/Matrix/vector with the
name(s) of the variable(s) that the protocols intervenes upon.
If only one treatment, then it would be one value with the name of treatment
"A" or the vector with one name for each time
A_0,A_1,----,A_k.  If more than one treatments, then it can be defined as a list
of length equal to the number of treatment
where each argument of the list has the vector of the treatment at different times
$A: A_0,A_1,...,A_k, $B:B_0,B_1,...B_k
OR it can be a matrix where each column contains the variables names
of treatment at different times.
\item \code{intervention}: A list/matrix/vector or a function.

If it is a matrix it should contain the values
that the variables are set to under the intervention as columns
and one row for each time point.
If the intervened values are the same for all time points it is sufficient to provide a single value per variable.
See examples. If it is a function, it will be called from \code{intervention_probabilities} with two arguments:
the current time interval and the current history of all variables. The function determines the value(s)
of the treatment variable(s) under the intervention and should return a matrix with as many columns as there are
treatment variables.
}}
}
\description{
This function adds a protocol to an existing object.
A protocol defines the values of the treatment variable(s)
at each time point during followup including at time zero (baseline).
}
