% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regLTMLE-package.R
\docType{package}
\name{regLTMLE-package}
\alias{regLTMLE}
\alias{regLTMLE-package}
\title{regLTMLE: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Thomas Alexander Gerds \email{tag@biostat.ku.dk} (\href{https://orcid.org/0000-0002-5955-816X}{ORCID})

Other contributors:
\itemize{
  \item Johan Sebastian Ohlendorff \email{johan.sebastian.ohlendorff@sund.ku.dk} [contributor]
}

}
\keyword{internal}
