% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cr_data.R
\name{sim_cr_data}
\alias{sim_cr_data}
\title{Simulate Competing risk Data}
\usage{
sim_cr_data(N, beta = NULL, eta = rep(0.1, 3), nu = rep(1.1, 3))
}
\arguments{
\item{N}{A double of the number of individuals}

\item{beta}{A 2X3 matrix with the effect of L0 and Treatment on the three processes.
The columns represent Process 1, Process 2 and Process 3, while the rows represent L0 and Treatment.}

\item{eta}{Vector of  length 3 of shape parameters for the Weibull hazard with parameterization
\deqn{\eta \nu t^{\nu - 1}}. Default is set to 0.1 for all events.}

\item{nu}{Vector of length 3 of scale parameters for the Weibull hazard.}
}
\value{
Data frame containing the simulated competing risk data. There is a column for ID, time of event (Time),
event type (Delta), baseline covariate (L0) and baseline Treatment (A0).
}
\description{
Function to simulate data from a Competing Risk setting. The function simulates data corresponding to $N$ individuals that
are at risk for mutually exclusive types of failure. 3 events can take place, one of which can be
interpreted as censoring.
}
\examples{
sim_cr_data(10)
}
