% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_setting1.R
\name{sim_data_setting1}
\alias{sim_data_setting1}
\title{Simulate Operation Data in Setting 1}
\usage{
sim_data_setting1(N, eta = rep(0.1, 4), nu = rep(1.1, 4), followup = Inf)
}
\arguments{
\item{N}{A double of the number of individuals}

\item{eta}{Vector of length 4 of shape parameters for the Weibull intensity with parameterization
\deqn{\eta \nu t^{\nu - 1}}. Default is set to 0.1 for all events.}

\item{nu}{Vector of scale parameters for the Weibull hazard. Default is set to 0.1 for all events.}

\item{followup}{A censoring time. By default set to infinity.}
}
\value{
Data frame containing the simulated data. There is a column for ID, time of event (Time),
event type (Delta), baseline covariate (L0), additional covariate (L) and  Treatment Process (A).
}
\description{
Function to simulate data from an Operation setting. 4 different types of events are simulated, chosen to represent
Operation (0),Death (1), Censoring(2) and Change in Covariate Process(3). Death and Censoring are terminal events
and Operation and Change in Covariate Process can occur once.
The intensities of the various events depend upon previous events and the pre specified \eqn{nu} and \eqn{eta}
parameters. The dependence on previous events is controlled by parameters chosen so that a large baseline covariate
(L0) increases the probability of operation(A = 1) and covariate change (L = 1). A change in the covariate process
(L= 1) increases the probability of operation (A = 1). Operation (A = 1)decreases the risk of death. A large baseline
covariate (L0), and a change in the covariate process(L = 1) increase the risk of death. The censoring process does
not depend on anything.
}
\examples{
sim_data_setting1(10)
}
