% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_op_data.R
\name{sim_op_data}
\alias{sim_op_data}
\title{Simulate Operation Data}
\usage{
sim_op_data(N, beta = NULL, eta = rep(0.1, 4), nu = rep(1.1, 4))
}
\arguments{
\item{N}{A double of the number of individuals}

\item{beta}{A 4X4 matrix of doubles for the effects on the intensities. The columns represent
the events Operation, Death, Censoring, and Covariate Change. The rows represent the baseline covariate \eqn{L0},
the indicator for operation \eqn{A}, baseline treatment \eqn{A0}, and the indicator for change in the covariate process \eqn{L}.
Default is set to 0.}

\item{eta}{Vector of length 4 of shape parameters for the Weibull intensity with parameterization
\deqn{\eta \nu t^{\nu - 1}}. Default is set to 0.1 for all events.}

\item{nu}{Vector of scale parameters for the Weibull hazard. Default is set to 0.1 for all events.}
}
\value{
Data frame containing the simulated data. There is a column for ID, time of event (Time),
event type (Delta), baseline covariate (L0), additional covariate (L), Baseline Treatment (A0) and
Treatment Process (A).
}
\description{
Function to simulate data from an Operation setting. 4 different types of events are simulated, chosen to represent
Operation (0),Death (1), Censoring(2) and Change in Covariate Process(3). Death and Censoring are terminal events
and Operation and Change in Covariate Process can occur once. The intensities of the various events depend upon
previous events and the pre specified \eqn{beta}, \eqn{nu} and \eqn{eta} parameters.
}
\examples{
sim_op_data(10)
}
