% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_recur_event_data.R
\name{sim_recur_event_data}
\alias{sim_recur_event_data}
\title{Simulate Recurrent Event Data}
\usage{
sim_recur_event_data(
  N,
  beta = NULL,
  eta = rep(0.1, 4),
  nu = rep(1.1, 4),
  at_risk = NULL,
  term_deltas = c(1, 2)
)
}
\arguments{
\item{N}{A double for the number of simulated individuals}

\item{beta}{A matrix of doubles for the effects of covariates on the hazard functions. The columns represent
the events operation, death, censoring, and covariate change. The rows represent the baseline covariate L0,
the event number k, treatment A, and the additional covariate L1. Default is set to 0.}

\item{eta}{Vector of shape parameters for the Weibull distribution. Default is set to 0.1.}

\item{nu}{Vector of scale parameters for the Weibull distribution. Default is set to 0.1.}

\item{at_risk}{At risk function. Default is set to recurrent event setting. A survival
or competing risk setting can be specified as well.}

\item{term_deltas}{Terminal events. Default is set so that event 1 and 2 are terminal events.}
}
\value{
Data frame containing the simulated data. There is a column for ID, time of event (Time),
event type (Delta), baseline covariate (L0), additional covariate (L1) and Treatment (A)
}
\description{
Function to simulate recurrent event data. The number of events simulated corresponds
to the length of the eta and nu vector, and the number of columns in the beta matrix.
Different settings can be specified: Survival setting, competing risk setting and operation setting.
}
\examples{
sim_recur_event_data(N = 10)
}
