% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete_functions.R
\name{target<-}
\alias{target<-}
\title{Define a target parameter}
\usage{
target(x, ...) <- value
}
\arguments{
\item{x}{An rtmle object as obtained with \code{rtmle_init}.}

\item{...}{Not (yet) used}

\item{value}{list with three forced elements:
\itemize{
\item \code{name}: the name of the target parameter.
\item \code{strategy}: the nuisance parameter modeling strategy consists of formulas and libraries for the regression models.
\item \code{estimator}: the estimator of the target parameter.
}}
}
\description{
A target defines a target parameter, an estimator and models
for the nuisance parameters.
}
