% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_int_data.R
\name{trans_int_data}
\alias{trans_int_data}
\title{Transformation of Event Data}
\usage{
trans_int_data(data)
}
\arguments{
\item{data}{Event data: a data frame containing an ID, Time, Delta, L0 and L column.}
}
\value{
Event data in a tstart tstop format
}
\description{
The function transforms the simulated event data containing time dependent covariates
into the tstart tstop format used by the coxph function.
}
\examples{
data <- sim_event_data(10)
trans_int_data(data)
}
