% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_sim_event_data.R
\name{vec_sim_event_data}
\alias{vec_sim_event_data}
\title{Simulate Event Data}
\usage{
vec_sim_event_data(
  N,
  beta = NULL,
  eta = rep(0.1, 4),
  nu = rep(1.1, 4),
  at_risk = NULL,
  term_deltas = c(1, 2)
)
}
\arguments{
\item{N}{A double for the number of simulated individuals}

\item{beta}{A matrix of doubles for the effects on the intensities. The columns represent
the events Operation, Death, Censoring, and Covariate Change. The rows represent the baseline covariate \eqn{L0},
the event number \eqn{k}, treatment \eqn{A}, and the indicator \eqn{L} for change in the covariate process.
Default is set to 0.}

\item{eta}{Vector of shape parameters for the Weibull hazard with parameterization
\deqn{\eta \nu t^{\nu - 1}}. Default is set to 0.1 for all events.}

\item{nu}{Vector of scale parameters for the Weibull hazard. Default is set to 0.1 for all events.}

\item{at_risk}{At risk function. Default is set to recurrent event setting. A survival
or competing risk setting can be specified as well. The \code{at_risk} function is an indicator of whether an individual is
at risk for a specific event. The function takes as input the event \eqn{x}, \eqn{k} and \eqn{L}.}

\item{term_deltas}{Terminal events. Default is set so that event 1 and 2 are terminal events.}
}
\value{
Data frame containing the simulated data. There is a column for ID, time of event (Time),
event type (Delta), baseline covariate (L0), additional covariate (L) and Treatment (A)
}
\description{
Vectorized function to simulate event data. The number of events simulated corresponds
to the length of the eta and nu vector, and the number of columns in the beta matrix.
By default 4 different types of events are simulated: Operation (0), Death (1), Censoring(2)
and Change in Covariate Process(4). Death and Censoring are terminal events and Operation
and Change in Covariate Process can occur once.
}
\details{
Different settings can be specified: Survival setting, competing risk setting and operation setting.
The different settings can be specified by the at_risk function. Default is the operation setting.
}
\examples{
#vec_sim_event_data(N = 10)
}
