% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_baseline_data.R
\name{add_baseline_data}
\alias{add_baseline_data}
\title{Adding baseline data to a rtmle object}
\usage{
add_baseline_data(x, data, ...)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{data}{data.frame with baseline data and id variable and a start_followup_date.}

\item{...}{Not used (not yet)}
}
\description{
This function adds a baseline dataset to an existing rtmle object
}
\examples{
x <- rtmle_init(time_grid = 0:2, name_id = "id", name_outcome = "Y")
baseline <- data.frame(id = 1:3,
                       age = c(55, 60, 65),
                       sex = c(0, 1, 0))
x <- add_baseline_data(x, baseline)
x$data$baseline_data
}
