% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_long_data.R
\name{add_long_data}
\alias{add_long_data}
\title{Adding long format data}
\usage{
add_long_data(
  x,
  outcome_data,
  censored_data,
  competing_data,
  timevar_data,
  ...
)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{outcome_data}{data.frame with two columns where the first is a subject identifier as initialized
under \code{x$names$id} and the second column contains dates of outcome events. 
Only subjects who exerience an outcome may have a row in the data.frame. The 
name of the date variable must match \code{x$names$time} and it may contain calendar time dates
or the pre-calculated time duration between time zero and the onset of the outcome.}

\item{censored_data}{data.frame with two columns where the first is a subject identifier as initialized
under \code{x$names$id} and the second column contains dates of end of followup.
All subjects can have a row in the data.frame. But the entries about the end of followup for people who experience the outcome or 
a competing risk are ignored.
The name of the date variable must match \code{x$names$time} and it may contain calendar time dates
or the pre-calculated time duration between time zero and the end of followup.}

\item{competing_data}{data.frame with two columns where the first is a subject identifier as initialized
under \code{x$names$id} and the second column contains dates of competing events. 
Only subjects who exerience an competing event may have a row in the data.frame. The 
name of the date variable must match \code{x$names$time} and it may contain calendar time dates
or the pre-calculated time duration between time zero and the onset of the competing event.}

\item{timevar_data}{named list of data.frames.}

\item{...}{Not used (not yet)}
}
\value{
The modified object.
}
\description{
Adding long format data to a rtmle object
}
\details{
This function adds a list of datasets with dates and values in long format (multiple lines per subject) to an existing rtmle object
}
\examples{
set.seed(17)
tau <- 3
ld <- simulate_long_data(n = 91,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(intervals = tau,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$outcome_data,
                   censored_data=ld$censored_data,
                   competing_data=ld$competing_data,
                   timevar_data=ld$timevar_data)
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,intervals = seq(0,2000,30.45*12))
x <- protocol(x,name = "Always_A",
                    intervention = data.frame("A" = factor("1",levels = c("0","1"))))
x <- protocol(x,name = "Never_A",
                    intervention = data.frame("A" = factor("0",levels = c("0","1"))))
x <- prepare_data(x)
x <- target(x,name = "Outcome_risk",
                  estimator = "tmle",
                  protocols = c("Always_A","Never_A"))
x <- model_formula(x)
x <- run_rtmle(x,learner = "learn_glmnet",time_horizon = 1:tau)
summary(x)
}
\seealso{
\link[rtmle]{add_baseline_data}, \link[rtmle]{add_wide_data}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
