% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wide_data.R
\name{add_wide_data}
\alias{add_wide_data}
\title{Add wide-format data}
\usage{
add_wide_data(x, outcome_data, timevar_data, ...)
}
\arguments{
\item{x}{An object of class \code{"rtmle"}, typically created by
\code{\link{rtmle_init}}.}

\item{outcome_data}{A wide-format data frame containing outcome variables and,
when present, censoring and competing-risk variables.}

\item{timevar_data}{A wide-format data frame, or a named list of data frames,
containing time-varying covariates.}

\item{...}{Not used.}
}
\value{
The modified \code{rtmle} object.
}
\description{
Add wide-format data to an rtmle object
}
\details{
Adds one-row-per-subject data in wide format to an existing \code{rtmle}
object.
}
\examples{
# Create wide-format data from simulated long-format data.
data(simulated_cohort)
ld <- register_format(simulated_cohort)
x <- rtmle_init(time_grid = seq(0,20,4), name_id = "id",
                name_outcome = "stroke", name_competing = "death",
                name_censoring = "dropout",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$timevar_data$stroke[!duplicated(id)],
                   censored_data=ld$timevar_data$dropout,
                   competing_data=ld$timevar_data$death,
                   timevar_data=ld$timevar_data[c("bleeding","changeSBP","A","B")])
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,start_followup_date=0)
outcome_data <- x$data$outcome_data
timevar_data <- x$data$timevar_data
x <- rtmle_init(time_grid = seq(0,20,4), name_id = "id",
                name_outcome = "stroke", name_competing = "death",
                name_censoring = "dropout",censored_label = "censored")
x <- add_baseline_data(x,data=ld$baseline_data)
x <- add_wide_data(x,outcome_data=outcome_data,timevar_data=timevar_data)

}
\seealso{
\code{\link{add_baseline_data}}, \code{\link{add_long_data}},
  \code{\link{long_to_wide}}, \code{\link{prepare_rtmle_data}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
