% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wide_data.R
\name{add_wide_data}
\alias{add_wide_data}
\title{Adding wide format data}
\usage{
add_wide_data(x, outcome_data, timevar_data, ...)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{outcome_data}{Data frame in a wide format. It should contain outcome variables,
possibly also censoring and/or competing risk variables, and time-varying covariates.}

\item{timevar_data}{Data frame in a wide format. It should contain the time-varying covariates.}

\item{...}{Not used (not yet)}
}
\value{
The modified object.
}
\description{
Adding wide format data to a rtmle object
}
\details{
This function adds a list of datasets in wide format (one line per subject) to an existing rtmle object
}
\examples{
# FIXME: this is a detour to produce wide format data
set.seed(112)
ld <- simulate_long_data(n = 11,number_visits = 20,
                         beta = list(A_on_Y = -.2,
                         A0_on_Y = -0.3,A0_on_A = 6),
                               register_format = TRUE)
x <- rtmle_init(intervals = 3, name_id = "id",
                name_outcome = "Y", name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$outcome_data,
                   censored_data=ld$censored_data,
                   competing_data=ld$competing_data,
                   timevar_data=ld$timevar_data)
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,intervals=seq(0,2000,30.45*6))
outcome_data <- x$data$outcome_data
timevar_data <- x$data$timevar_data
x <- rtmle_init(intervals = 3, name_id = "id",
                name_outcome = "Y", name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_baseline_data(x,data=ld$baseline_data)
x <- add_wide_data(x,outcome_data=outcome_data,timevar_data=timevar_data)
x <- prepare_data(x)

}
\seealso{
\link[rtmle]{add_baseline_data}, \link[rtmle]{add_long_data}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
