% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheap_bootstrap.R
\name{cheap_bootstrap}
\alias{cheap_bootstrap}
\title{Cheap subsampling bootstrap}
\usage{
cheap_bootstrap(
  x,
  B = 20,
  M,
  replace = FALSE,
  seeds,
  alpha = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{B}{Number of bootstrap samples.}

\item{M}{Size of the bootstrap samples. If argument \code{replace}
is \code{FALSE} this defaults to \code{0.632 *
    NROW(x$prepared_data)} and if argument \code{replace} is
\code{TRUE} this defaults to \code{NROW(x$prepared_data)}.}

\item{replace}{Logical. Default is \code{FALSE}. If \code{FALSE}
bootstrap samples are obtained by sampling without
replacement. In this case the value of argument \code{M} must
be smaller than \code{NROW(x$prepared_data)}.  # If \code{TRUE}
bootstrap samples are obtained by sampling with replacement. In
this case argument, by default \code{M} will be set equal to
\code{NROW(x$prepared_data)} # Defaults to "subsampling".}

\item{seeds}{A vector of random seeds for controlling the
randomness while drawing the bootstrap samples.}

\item{alpha}{Significance level. Defaults to 0.05.}

\item{verbose}{Logical. Passed to run_rtmle to control verbosity.}
}
\value{
The modified object
}
\description{
Method implementing the cheap subsampling method for confidence intervals
}
\details{
Given a model object or a function that returns a vector of coefficients
and a data set,
this function computes confidence intervals using
the cheap subsampling method.
}
\examples{
\dontrun{
set.seed(17)
tau <- 3
ld <- simulate_long_data(n = 100,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(intervals = tau,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x$long_data <- ld[c("outcome_data","censored_data","competing_data","timevar_data")]
add_baseline_data(x) <- ld$baseline_data[,start_followup_date:=0]
x <- long_to_wide(x,intervals = seq(0,2000,30.45*12))
protocol(x) <- list(name = "Always_A",
                    intervention = data.frame("A" = factor("1",levels = c("0","1"))))
protocol(x) <- list(name = "Never_A",
                    intervention = data.frame("A" = factor("0",levels = c("0","1"))))
prepare_data(x) <- list()
target(x) <- list(name = "Outcome_risk",
                  strategy = "additive",
                  estimator = "tmle",
                  protocols = c("Always_A","Never_A"))
 x <- run_rtmle(x,learner = "learn_glm",time_horizon = 1:tau)
 x <- cheap_bootstrap(x,B=5,time_horizon=1:tau)
 summary(x)
cb
}

}
\author{
Johan Sebastian Ohlendorff
\email{johan.ohlendorff@sund.ku.dk} and Thomas A Gerds
\email{tag@biostat.ku.dk}
}
