% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheap_bootstrap.R
\name{cheap_bootstrap}
\alias{cheap_bootstrap}
\title{Cheap subsampling bootstrap}
\usage{
cheap_bootstrap(
  x,
  time_horizon,
  B = 20,
  M,
  seeds,
  alpha = 0.05,
  add = TRUE,
  verbose = FALSE,
  which_subsets = NULL
)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{time_horizon}{The time horizon at which to calculate
the bootstrap confidence intervals. If it is a vector the analysis will be performed for
each element of the vector.}

\item{B}{Number of bootstrap samples.}

\item{M}{Size of the bootstrap samples. The defaults is \code{0.632 * NROW(x$prepared_data)}.}

\item{seeds}{A vector of random seeds for controlling the
randomness while drawing the bootstrap samples.}

\item{alpha}{Significance level. Defaults to 0.05.}

\item{add}{Logical. If \code{TRUE} add new bootstrap results to the existing bootstrap results}

\item{verbose}{Logical. Passed to run_rtmle to control verbosity.}

\item{which_subsets}{Character vector specifying which subsets to perform the cheap bootstrap on.}
}
\value{
The modified object
}
\description{
Method implementing the cheap subsampling method for confidence intervals
}
\details{
Given a model object or a function that returns a vector of coefficients
and a data set,
this function computes confidence intervals using
the cheap subsampling method.
}
\examples{
\dontrun{
set.seed(17)
tau <- 3
ld <- simulate_long_data(n = 100,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(intervals = tau,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$outcome_data,
                   censored_data=ld$censored_data,
                   competing_data=ld$competing_data,
                   timevar_data=ld$timevar_data)
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,intervals = seq(0,2000,30.45*12))
x <- protocol(x,name = "Always_A",
                    intervention = data.frame("A" = factor("1",levels = c("0","1"))))
x <- protocol(x,name = "Never_A",
                    intervention = data.frame("A" = factor("0",levels = c("0","1"))))
x <- prepare_data(x)
x <- target(x,name = "Outcome_risk",
                  estimator = "tmle",
                  protocols = c("Always_A","Never_A"))
 x <- model_formula(x)
 x <- run_rtmle(x,learner = "learn_glm",time_horizon = 1:tau)
 x <- cheap_bootstrap(x,B=5,time_horizon=1:tau, M = 0.632*NROW(x$prepared_data))
 summary(x)
}

}
\author{
Johan Sebastian Ohlendorff
    \email{johan.ohlendorff@sund.ku.dk} and Thomas A Gerds
    \email{tag@biostat.ku.dk}
}
