% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheap_bootstrap.R
\name{cheap_bootstrap}
\alias{cheap_bootstrap}
\title{Cheap subsampling}
\usage{
cheap_bootstrap(
  fun,
  data,
  size = NULL,
  type = "subsampling",
  b = 20,
  estimate_column_name = "estimate",
  parameter_column_names = "parameter",
  alpha = 0.05,
  keep_estimates = TRUE
)
}
\arguments{
\item{fun}{A function that returns a data.frame
in estimate_column_name and the corresponding parameter names in parameter_column_names.}

\item{data}{Data set to be used for the computation.}

\item{size}{Subsample size. Defaults to 0.632 * nrow(data).}

\item{type}{Character. Type of bootstrap method.
Can be either "subsampling" or "non_parametric".
Defaults to "subsampling".}

\item{b}{Number of bootstrap samples.}

\item{estimate_column_name}{Character. Column name of the point estimates.}

\item{parameter_column_names}{Character. Column names of the parameters.}

\item{alpha}{Significance level. Defaults to 0.05.}

\item{keep_estimates}{Logical. Should the bootstrapped estimates be kept?}
}
\value{
An object of class "cheap_bootstrap" containing
the point estimates and confidence intervals.
}
\description{
Method implementing the cheap subsampling method for confidence intervals
}
\details{
Given a model object or a function that returns a vector of coefficients
and a data set,
this function computes confidence intervals using
the cheap subsampling method.
}
\examples{
utils::data(anorexia, package = "MASS")
## example with a function call
set.seed(123)
library(broom)
library(dplyr)
fun <- function(data) {
  lm(formula = Postwt ~ Prewt + Treat + offset(Prewt), data = data) \%>\%
  tidy()
}
cs <- cheap_bootstrap(fun = fun,
                      b = 20,
                      data = anorexia,
                      estimate_column_name = "estimate",
                      parameter_column_names = "term")
cs
summary(cs)
plot(cs)

}
