% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_net.R
\name{glm_net}
\alias{glm_net}
\title{Fitting glm_net for use with rtmle}
\usage{
glm_net(
  formula,
  data,
  lambda = NULL,
  cv = TRUE,
  alpha = 1,
  nfolds = 10,
  type.measure = "deviance",
  family,
  ...
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{The data on which to fit the model.}

\item{lambda}{The tuning parameters for glm_net. If set to NULL, then it the parameters are chosen for you.}

\item{cv}{Whether to use cross-validation or not. Default is TRUE.}

\item{alpha}{The elasticnet mixing parameter. See the ?glmnet for more details.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{type.measure}{loss to use for cross-validation. Default is deviance.}

\item{family}{passed to \code{glmnet}. Defaults for binary outcome to \code{"binomial"} and for survival to \code{"cox"}.}

\item{\dots}{Additional arguments that are passed on to the glmnet.}
}
\description{
Fit glm_net models via a formula and a data set for use with learn_glmnet
}
