% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_glm.R
\name{learn_glm}
\alias{learn_glm}
\title{Nuisance parameter learner based on \code{\link{glm}}}
\usage{
learn_glm(
  character_formula,
  data,
  intervened_data,
  learn_variables = NULL,
  ...
)
}
\arguments{
\item{character_formula}{Formula for nuisance parameter as a character}

\item{data}{Data for learning}

\item{intervened_data}{Data for prediction}

\item{learn_variables}{Vector of variable names. Only include these in the learning.
Can be \code{"NONE"} which means to not use any variables and simply predict the mean outcome.}

\item{...}{Additional arguments for the learning phase. Not used at the moment.}
}
\value{
A vector of predicted probabilities which has the fit as an attribute.
}
\description{
Learning nuisance parameter models for TMLE and predicting
probabilities in intervened data based on \code{\link{glm}}
}
\details{
The function attempts to run \code{speedglm} and uses \code{glm} only as a fallback option.
Optional argument learn_variables can be used to restrict the learning to a subset of variables.
This can be useful to avoid too many parameters in the model.
}
\seealso{
\code{link{superlearn}}, \code{link{learn_ranger}}, \code{link{learn_glmnet}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
