% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_glmnet.R
\name{learn_glmnet}
\alias{learn_glmnet}
\title{Nuisance parameter learner based on \code{\link[glmnet]{glmnet}}}
\usage{
learn_glmnet(
  character_formula,
  data,
  intervened_data,
  selector = "undersmooth",
  lambda = NULL,
  alpha = 1,
  family,
  nfolds = 10,
  type.measure = "deviance",
  ...
)
}
\arguments{
\item{character_formula}{Formula for nuisance parameter as a character}

\item{data}{Data for learning}

\item{intervened_data}{Data for prediction}

\item{selector}{Character value deciding about how to select the penalty parameter lambda:
If \code{"undersmooth"} use the lambda value which results in the least amount of
penalty such that the model still fits. The other options are \code{"min"} and \code{"1se"} which are 
described in the documentation of \link[glmnet]{cv.glmnet}.}

\item{lambda}{Penalty parameter passed to \link[glmnet]{glmnet} or \link[glmnet]{cv.glmnet}.}

\item{alpha}{The elasticnet mixing parameter. See \link[glmnet]{glmnet} for more details.}

\item{family}{Passed to \link[glmnet]{glmnet} or \link[glmnet]{cv.glmnet}.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{type.measure}{loss to use for cross-validation. Default is deviance.}

\item{...}{Additional arguments for the learning phase passed to \code{\link[glmnet]{glmnet}}. 
E.g., setting alpha affects the elastic net.}
}
\value{
A vector of predicted probabilities which has the fit as an attribute.
}
\description{
Learning nuisance parameter models for TMLE and predicting
probabilities in intervened data based on \code{\link[glmnet]{glmnet}}
}
\details{
This can be useful to avoid too many parameters in the model.
}
\examples{
d <- data.table::data.table(Y = rep(c(0, 1), 10),
                            A = rep(c(0, 1, 1, 0), 5),
                            L = seq(-1, 1, length.out = 20))
predicted <- learn_glmnet("Y ~ A + L", data = d, intervened_data = d,
                          lambda = 0.01)
head(predicted)
attr(predicted, "selected.lambda")
}
\seealso{
\code{link{superlearn}}, \code{link{learn_ranger}}, \code{link{learn_glm}}
}
