% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_glmnet.R
\name{learn_glmnet}
\alias{learn_glmnet}
\title{Nuisance parameter learner based on \code{\link[glmnet]{glmnet}}}
\usage{
learn_glmnet(
  character_formula,
  data,
  intervened_data,
  selector = "undersmooth",
  ...
)
}
\arguments{
\item{character_formula}{Formula for nuisance parameter as a character}

\item{data}{Data for learning}

\item{intervened_data}{Data for prediction}

\item{selector}{Character value deciding about how to select the penalty parameter lambda:
If \code{"undersmooth"} use the lambda value which results in the least amount of
penalty such that the model still fits. The other options are \code{"min"} and \code{"1se"} which are 
described in the documentation of \code{\link[glmnet]{cv.glmnet}}.}

\item{...}{Additional arguments for the learning phase passed to \code{\link[glmnet]{glmnet}}. 
E.g., setting alpha affects the elastic net.}
}
\value{
A vector of predicted probabilities which has the fit as an attribute.
}
\description{
Learning nuisance parameter models for TMLE and predicting
probabilities in intervened data based on \code{\link[glmnet]{glmnet}}
}
\details{
This can be useful to avoid too many parameters in the model.
}
\seealso{
\code{link{superlearn}}, \code{link{learn_ranger}}, \code{link{learn_glm}}
}
