% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_ranger.R
\name{learn_ranger}
\alias{learn_ranger}
\title{Nuisance parameter learner based on \code{\link[ranger]{ranger}}}
\usage{
learn_ranger(character_formula, data, intervened_data, ...)
}
\arguments{
\item{character_formula}{Formula for nuisance parameter as a character}

\item{data}{Data for learning}

\item{intervened_data}{Data for prediction}

\item{...}{Additional arguments for the learning phase passed to \code{\link[ranger]{ranger}}. These can include hyperparameters.}
}
\value{
A vector of predicted probabilities which has the fit as an attribute.
}
\description{
Learning nuisance parameter models for TMLE and predicting
probabilities in intervened data based on \code{\link[ranger]{ranger}}
}
\details{
Hyperparameters are set via ...
}
\examples{
d <- data.table::data.table(Y = factor(rep(c(0, 1), 10)),
                            A = rep(c(0, 1, 1, 0), 5),
                            L = seq(-1, 1, length.out = 20))
predicted <- learn_ranger("Y ~ A + L", data = d, intervened_data = d,
                          num.trees = 10, min.node.size = 1)
head(predicted)
}
\seealso{
\code{link{superlearn}}, \code{link{learn_glm}}, \code{link{learn_glmnet}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
