% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_ranger.R
\name{learn_ranger}
\alias{learn_ranger}
\title{Nuisance-parameter learner based on ranger}
\usage{
learn_ranger(character_formula, data, intervened_data, ...)
}
\arguments{
\item{character_formula}{Formula for the nuisance parameter, supplied as a
character string.}

\item{data}{Data used for learning.}

\item{intervened_data}{Data used for prediction after intervention variables
have been set according to a protocol.}

\item{...}{Additional arguments passed to \code{\link[ranger]{ranger}},
including hyperparameters.}
}
\value{
A vector of predicted probabilities with the fitted model stored in
  the \code{"fit"} attribute.
}
\description{
Learn nuisance-parameter models with ranger
}
\details{
Learns nuisance-parameter models for TMLE and predicts probabilities in
intervention-updated data using \code{\link[ranger]{ranger}}.
}
\examples{
d <- data.table::data.table(Y = factor(rep(c(0, 1), 10)),
                            A = rep(c(0, 1, 1, 0), 5),
                            L = seq(-1, 1, length.out = 20))
if (requireNamespace("ranger", quietly = TRUE)) {
predicted <- learn_ranger("Y ~ A + L", data = d, intervened_data = d,
                          num.trees = 10, min.node.size = 1)
head(predicted)
}
}
\seealso{
\code{\link{superlearn}}, \code{\link{learn_glm}},
  \code{\link{learn_glmnet}}, \code{\link{learn_xgboost}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
