% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_xgboost.R
\name{learn_xgboost}
\alias{learn_xgboost}
\title{Nuisance-parameter learner based on xgboost}
\usage{
learn_xgboost(
  character_formula,
  data,
  intervened_data,
  save_fitted_objects = FALSE,
  reuse_fit = NULL,
  ...
)
}
\arguments{
\item{character_formula}{Formula for the nuisance parameter, supplied as a
character string.}

\item{data}{Data used for learning.}

\item{intervened_data}{Data used for prediction after intervention variables
have been set according to a regime.}

\item{save_fitted_objects}{Logical. If \code{TRUE}, store the 
fitted object as element \code{fit}}

\item{reuse_fit}{Optional fitted object returned by a previous call. If
supplied, skip fitting and only predict in \code{intervened_data}.}

\item{...}{Additional arguments passed to \code{\link[xgboost]{xgboost}},
including hyperparameters.}
}
\value{
A list whose first element, \code{predicted_values}, is a vector of
  predicted probabilities. Element \code{object} contains the fitted model.
}
\description{
Learn nuisance-parameter models with xgboost
}
\details{
Learns nuisance-parameter models for TMLE and predicts probabilities in
intervention-updated data using \code{\link[xgboost]{xgboost}}.
}
\examples{
if (requireNamespace("xgboost", quietly = TRUE)) {
tau <- 3
data(simulated_cohort)
ld <- register_format(simulated_cohort)
x <- rtmle_init(time_grid = seq(0,20,4),name_id = "id",
                name_outcome = "stroke",name_competing = "death",
                name_censoring = "dropout",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$timevar_data$stroke[!duplicated(id)],
                   censored_data=ld$timevar_data$dropout,
                   competing_data=ld$timevar_data$death,
                   timevar_data=ld$timevar_data[c("bleeding","changeSBP","A","B")])
x <- add_baseline_data(x,data=ld$baseline_data)
x <- discretize_data(x,start_followup_date=0)
x <- prepare_rtmle_data(x)
x <- regime(x,name = "Always_A",
                    intervention = data.frame(time=x$intervention_nodes,
                                              "A" = factor("1",levels = c("0","1"))))
x <- regime(x,name = "Never_A",
                    intervention = data.frame(time=x$intervention_nodes,
                                              "A" = factor("0",levels = c("0","1"))))
x <- target(x,name = "Outcome_risk",
                  estimator = "tmle",
                  regimes = c("Always_A","Never_A"))
x <- model_formula(x)
x <- run_rtmle(x,learner = "learn_xgboost",time_horizon = 3)
}
}
\seealso{
\code{\link{superlearn}}, \code{\link{learn_glm}},
  \code{\link{learn_glmnet}}, \code{\link{learn_ranger}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
