% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_xgboost.R
\name{learn_xgboost}
\alias{learn_xgboost}
\title{Nuisance-parameter learner based on xgboost}
\usage{
learn_xgboost(character_formula, data, intervened_data, ...)
}
\arguments{
\item{character_formula}{Formula for the nuisance parameter, supplied as a
character string.}

\item{data}{Data used for learning.}

\item{intervened_data}{Data used for prediction after intervention variables
have been set according to a protocol.}

\item{...}{Additional arguments passed to \code{\link[xgboost]{xgboost}},
including hyperparameters.}
}
\value{
A vector of predicted probabilities with the fitted model stored in
  the \code{"fit"} attribute.
}
\description{
Learn nuisance-parameter models with xgboost
}
\details{
Learns nuisance-parameter models for TMLE and predicts probabilities in
intervention-updated data using \code{\link[xgboost]{xgboost}}.
}
\examples{
set.seed(17)
tau <- 3
ld <- simulate_long_data(n = 91,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(time_grid = seq(0,1500,30.45*12),name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$outcome_data,
                   censored_data=ld$censored_data,
                   competing_data=ld$competing_data,
                   timevar_data=ld$timevar_data)
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,start_followup_date=0)
x <- protocol(x,name = "Always_A",
                    intervention = data.frame(time=x$intervention_nodes,
                                              "A" = factor("1",levels = c("0","1"))))
x <- protocol(x,name = "Never_A",
                    intervention = data.frame(time=x$intervention_nodes,
                                              "A" = factor("0",levels = c("0","1"))))
x <- prepare_rtmle_data(x)
x <- target(x,name = "Outcome_risk",
                  estimator = "tmle",
                  protocols = c("Always_A","Never_A"))
x <- model_formula(x)
x <- run_rtmle(x,learner = "learn_xgboost",time_horizon = 3)
}
\seealso{
\code{\link{superlearn}}, \code{\link{learn_glm}},
  \code{\link{learn_glmnet}}, \code{\link{learn_ranger}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
