% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_xgboost.R
\name{learn_xgboost}
\alias{learn_xgboost}
\title{Nuisance parameter learner based on \code{\link[xgboost]{xgboost}}}
\usage{
learn_xgboost(character_formula, data, intervened_data, ...)
}
\arguments{
\item{character_formula}{Formula for nuisance parameter as a character}

\item{data}{Data for learning}

\item{intervened_data}{Data for prediction}

\item{...}{Additional arguments for the learning phase passed to \code{\link[xgboost]{xgboost}}. These can include hyperparameters.}
}
\value{
A vector of predicted probabilities which has the fit as an attribute.
}
\description{
Learning nuisance parameter models for TMLE and predicting
probabilities in intervened data based on \code{\link[xgboost]{xgboost}}
}
\details{
Hyperparameters are set via ...
}
\examples{
set.seed(17)
tau <- 3
ld <- simulate_long_data(n = 391,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(intervals = tau,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$outcome_data,
                   censored_data=ld$censored_data,
                   competing_data=ld$competing_data,
                   timevar_data=ld$timevar_data)
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,intervals = seq(0,2000,30.45*12))
x <- protocol(x,name = "Always_A",
                    intervention = data.frame("A" = factor("1",levels = c("0","1"))))
x <- protocol(x,name = "Never_A",
                    intervention = data.frame("A" = factor("0",levels = c("0","1"))))
x <- prepare_data(x)
x <- target(x,name = "Outcome_risk",
                  estimator = "tmle",
                  protocols = c("Always_A","Never_A"))
x <- model_formula(x)
x <- run_rtmle(x,learner = "learn_xgboost",time_horizon = 3)
}
\seealso{
\code{link{superlearn}}, \code{link{learn_glm}}, \code{link{learn_glmnet}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
