% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_wide.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Transform data from long to wide format}
\usage{
long_to_wide(
  x,
  intervals,
  fun = function(x) {
     1 * (sum(x) > 0)
 }
)
}
\arguments{
\item{x}{object as obtained with rtmle_init}

\item{intervals}{a vector of time points that discretize the followup time into intervals}

\item{fun}{function used to map information onto the intervals. see Details.}
}
\description{
This function is used to prepare the discretized time analysis of register data.
The start_followup_date is the calendar date where the follow up starts
or can be zero if the followup data are readily on the time on study scale
}
\details{
TODO
}
