% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exclusion_rule.R
\name{make_exclusion_rule}
\alias{make_exclusion_rule}
\title{Helper function for model formula}
\usage{
make_exclusion_rule(
  timepoint,
  name_outcome_variable,
  excluded_variables,
  recursive = FALSE
)
}
\arguments{
\item{timepoint}{Maximal time point of interest}

\item{name_outcome_variable}{Name of time-dependent outcome/treatment/censoring variable without suffix.}

\item{excluded_variables}{Vector of time-dependent variables (without suffix) to be excluded.}

\item{recursive}{Logical. If \code{TRUE} exclude all historical variables otherwise only the one from the same time point.}
}
\value{
Named list
}
\description{
Helper function for model_formula
}
\details{
This function creates a list of variables that should be excluded
from the formulas of a sequence of
}
\examples{
make_exclusion_rule(4,"A","B",recursive = TRUE)
make_exclusion_rule(4,"A",c("B","C"),recursive = TRUE)
}
\seealso{
model_formula
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
