% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_formula.R
\name{model_formula}
\alias{model_formula}
\title{Model formulas for nuisance parameters}
\usage{
model_formula(
  x,
  propensity_model = "joint",
  Markov = NULL,
  verbose = TRUE,
  exclude_variables = NULL,
  exclusion_rules = NULL,
  inclusion_rules = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"rtmle"} with prepared data and at least
one protocol defined by \code{\link{protocol}}.}

\item{propensity_model}{Relevant only for interventions depending on
multiple treatment variables. Controls how to model the propensity of the
regimen dictated by the protocol. Possible values are \code{"joint"},
\code{"sequential"}, and \code{"independent"}.}

\item{Markov}{Names of time-dependent variables that should appear only with
their most recent values on the right-hand side of formulas.}

\item{verbose}{Logical. If \code{FALSE} suppress all messages. \code{TRUE} is the default.}

\item{exclude_variables}{Variables to exclude from the formulas for the nuisance parameters.}

\item{exclusion_rules}{Experimental. Additional exclusion rules given as a
named list. Names are variables occurring on the left-hand side of a
formula, and elements are variables to exclude from the right-hand side.}

\item{inclusion_rules}{Experimental. Additional inclusion rules given as a
named list. Names are variables occurring on the left-hand side of a
formula, and elements are variables to include on the right-hand side.}

\item{...}{Not used.}
}
\value{
The modified \code{rtmle} object.
}
\description{
Specify formulas for nuisance-parameter models
}
\details{
Because the data are discretized, time-dependent covariates at time \code{k}
are not included in treatment models at time \code{k}, except at time 0.
This avoids encoding an ordering assumption in which \code{L_k} precedes
\code{A_k} when, in the original event history, \code{A_k} may occur before
\code{L_k}.
}
\examples{
data(simulated_cohort)
ld <- register_format(simulated_cohort)
x <- rtmle_init(time_grid = seq(0,20,4),name_id = "id",name_outcome = "stroke",
                name_competing = "death",
                name_censoring = "dropout",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$timevar_data$stroke[!duplicated(id)],
                   censored_data=ld$timevar_data$dropout,
                   competing_data=ld$timevar_data$death,
                   timevar_data=ld$timevar_data[c("bleeding","changeSBP","A","B")])
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,start_followup_date=0)
x <- prepare_rtmle_data(x)
x <- protocol(x,name = "Always_A",
              intervention = data.frame(time=x$intervention_nodes,
                                        "A" = factor("1",levels = c("0","1"))))
x <- protocol(x,name = "Never_A",
              intervention = data.frame(time=x$intervention_nodes,
                                        "A" = factor("0",levels = c("0","1"))))
x <- protocol(x,name = "Use_A_not_B",
              intervention = data.frame(time=x$intervention_nodes,
                                        "A" = factor("1",levels = c("0","1")),
                                        "B" = factor("0",levels = c("0","1"))))
x <- model_formula(x)
x$models
# remove age from all formulas
x <- model_formula(x,exclusion_rules=list("*"="age"))
# remove age from the stroke_1 formula
x <- model_formula(x,exclusion_rules=list("stroke_1"="age"))
x$models
# remove age from all stroke formulas
x <- model_formula(x,exclusion_rules=list("stroke_*"="age"))
# remove age and changeSBP_0 from dropout_1 and dropout_2 formulas
x <- model_formula(x,exclusion_rules=list("dropout_[1:2]"="age|changeSBP_0"))
x$models
# remove all changeSBP_t variables from dropout_1 and dropout_2 formulas
x <- model_formula(x,exclusion_rules=list("dropout_[1:2]"="changeSBP_*"))
x$models
}
\seealso{
\code{\link{prepare_rtmle_data}}, \code{\link{protocol}},
  \code{\link{target}}, \code{\link{make_exclusion_rule}},
  \code{\link{run_rtmle}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
