% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_formula.R
\name{model_formula}
\alias{model_formula}
\title{Model formulas for nuisance parameters}
\usage{
model_formula(
  x,
  Markov = NULL,
  verbose = TRUE,
  exclude_variables = NULL,
  exclusion_rules = NULL,
  inclusion_rules = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{Markov}{Names of time-dependent variables which should only occur with the most recent
value on the right hand side of the formulas.}

\item{verbose}{Logical. If \code{FALSE} suppress all messages. \code{TRUE} is the default.}

\item{exclude_variables}{Variables to exclude from the formulas for the nuisance parameters.}

\item{exclusion_rules}{Experimental. Additional exclusion rules given as a named list where names are variables that occur on the left hand side of a formula and
elements are variables that should be included in the right hand side of the formula.}

\item{inclusion_rules}{Experimental. Additional inclusion rules given as a named list where names are variables that occur on the left hand side of a formula and
elements are variables that should be included in the right hand side of the formula.}

\item{...}{Not used (not yet)}
}
\value{
The modified \code{rtmle}object
}
\description{
Specify model formulas for the estimators of the nuisance parameters
}
\details{
Note that due to the discretized nature of the data, we do not include
the time-dependent covariates at time k in the models for treatment at time k,
except at time 0. 
The reason for this is we do not want to mistakenly assume that L_k -> A_k when in reality A_k may happen before L_k
}
\examples{
set.seed(112)
ld <- simulate_long_data(n = 1000,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(intervals = 2,name_id = "id",name_outcome = "Y",
                name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_long_data(x,outcome_data = ld[["outcome_data"]],
                     censored_data = ld[["censored_data"]],
                     competing_data = ld[["competing_data"]],
                     timevar_data = ld[["timevar_data"]])
x <- add_baseline_data(x,data = ld$baseline_data)
x <- long_to_wide(x,intervals = seq(0,2000,30.45*6))
x <- protocol(x,name = "Always_A",treatment_variables = "A",intervention = 1)
x <- prepare_data(x) 
x <- target(x,name = "Outcome_risk",estimator = "tmle",protocols = "Always_A")
x <- model_formula(x)
x$models
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
