% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_learners.R
\name{parse_learners}
\alias{parse_learners}
\title{Parsing list of learners}
\usage{
parse_learners(learners)
}
\arguments{
\item{learners}{List of learners}
}
\value{
List of learners
}
\description{
Parsing list of learners for function superlearn. Available names are 
substituted for missing learner_fun elements and available elements learner_fun
are substituted for missing names.
}
\details{
As a side effect an attribute is given to the object such that when running
on its own output the function does nothing.
}
\examples{
parse_learners(c("learn_glm","learn_glm"))
parse_learners(c(list("learn_glmnet",
                 "glm"=list(learn_variables="A")),list("learn_ranger"),
                 list(list(learner_fun="learn_ranger",num.trees=5))))
}
\seealso{
\code{\link{superlearn}}, \code{\link{run_rtmle}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
