% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rtmle.R
\name{plot.rtmle}
\alias{plot.rtmle}
\alias{autoplot.rtmle}
\title{Plot rtmle objects}
\usage{
\method{autoplot}{rtmle}(
  object,
  analysis = NULL,
  xlim,
  ylim,
  y_breaks,
  x_breaks,
  position_atrisk,
  conf_int,
  ...
)

\method{plot}{rtmle}(x, ...)
}
\arguments{
\item{object}{A prepared and fitted object of class \code{"rtmle"}.}

\item{analysis}{Name of the analysis. If \code{NULL}, use
\code{object$estimate[["Main_analysis"]]}.}

\item{xlim}{Limits for the x-axis.}

\item{ylim}{Limits for the y-axis.}

\item{y_breaks}{Breaks for the y-axis.}

\item{x_breaks}{Breaks for the x-axis.}

\item{position_atrisk}{Vector of positions on the x-axis where numbers at-risk are shown below the graph.}

\item{conf_int}{Logical. If \code{TRUE}, add confidence bands to the graph.}

\item{...}{Not used.}

\item{x}{A prepared and fitted object of class \code{"rtmle"}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot risk estimates and risk contrasts from fitted
  \code{rtmle} objects.
}
\examples{
x <- list(
  estimate = list(Main_analysis = data.table::data.table(
    Time_horizon = c(1, 2, 1, 2),
    Time = c(1, 2, 1, 2),
    Protocol = rep(c("Always_A", "Never_A"), each = 2),
    Estimate = c(.12, .20, .18, .30),
    Lower = c(.08, .15, .12, .24),
    Upper = c(.18, .27, .25, .38))),
  followup = data.table::data.table(id = 1:5,
                                    last_interval = c(0, 1, 2, 2, 2)),
  time_grid_scale = 0:2)
class(x) <- "rtmle"
p <- ggplot2::autoplot(x, xlim = c(0, 2))
class(p)
}
\seealso{
\code{\link{run_rtmle}}, \code{\link{summary.rtmle}},
  \code{\link{plot_adherence}}, \code{\link{plot_IPW}}
}
