% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rtmle.R
\name{plot.rtmle}
\alias{plot.rtmle}
\alias{autoplot.rtmle}
\title{ggplot rtmle objects}
\usage{
\method{autoplot}{rtmle}(
  object,
  analysis = NULL,
  xlim,
  ylim,
  y_breaks,
  x_breaks,
  position_atrisk,
  conf_int,
  ...
)

\method{plot}{rtmle}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"rtmle"} prepared and fitted}

\item{analysis}{Name of the analysis. Default is NULL which refers to object[["Main_analysis"]].}

\item{xlim}{Limits for x-axis}

\item{ylim}{Limits for y-axis}

\item{y_breaks}{Breaks for y-axis}

\item{x_breaks}{Breaks for x-axis}

\item{position_atrisk}{Vector of positions on the x-axis where numbers at-risk are shown below the graph.}

\item{conf_int}{Logical. If \code{TRUE} add confidence shadows to the graph.}

\item{...}{Not used}

\item{x}{Object of class \code{"rtmle"} prepared and fitted}
}
\description{
Plot risk predictions and average treatment effects from rtmle objects
}
