% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare}
\alias{plot_compare}
\title{Plot comparison of estimates.}
\usage{
plot_compare(
  estimates,
  diff_betas = seq(0, 1, by = 0.1),
  lower = NULL,
  upper = NULL,
  groups = c(0, -0.1, -0.2)
)
}
\arguments{
\item{estimates}{A matrix with estimates. Columns 1,3,... are for A0 = 0 and
columns 2,4,... for  A0 = 1}

\item{diff_betas}{The values of the varying beta coefficient.}

\item{lower}{A matrix with lower confidence bands. Columns 1,3,... are for A0 = 0 and
columns 2,4,... for  A0 = 1}

\item{upper}{A matrix with lower confidence bands. Columns 1,3,... are for A0 = 0 and
columns 2,4,... for  A0 = 1}

\item{groups}{The strata for the different estimates}
}
\value{
Plot
}
\description{
\code{plot_compare} is a function for plotting an estimates. It takes a matrix of estimates, and plots.
}
\examples{
# First we define an estimator
estimator1 <- function(data, N) {

T2D_events <- data[Delta == 3]
return(c(nrow(T2D_events[A0 == 1 & Time < 5])/N, nrow(T2D_events[A0 == 0 & Time < 5])/N))

}

res1 <- compare_effects(estimator = estimator1, N = 100, beta_L_D = seq(0,1,by = 0.1))
plot_compare(res1, diff_betas = seq(0,1,by = 0.1))
}
