% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_event_data.R
\name{plot_event_data}
\alias{plot_event_data}
\title{Plot Event Data}
\usage{
plot_event_data(data, title = "Event Data")
}
\arguments{
\item{data}{Event data: a data frame containing an ID, Time, Delta, L0 and L column.}

\item{title}{Title: a string that will be the title of the plot.}
}
\value{
A plot of the event data
}
\description{
Function to plot simulated event data.
}
\examples{
data <- sim_event_data(10)
plot_event_data(data)
}
