% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rtmle_data.R
\name{prepare_rtmle_data}
\alias{prepare_rtmle_data}
\title{Prepare a targeted minimum loss-based analysis}
\usage{
prepare_rtmle_data(x, ...)
}
\arguments{
\item{x}{An object initialized with \code{\link{rtmle_init}} containing
wide-format data, usually added with \code{\link{add_wide_data}} or
created from long-format data with \code{\link{discretize}}.}

\item{...}{Not used.}
}
\value{
The object augmented with a new element called
  \code{prepared_data}.
}
\description{
Prepare an \code{rtmle} object for estimation by checking the supplied data
and storing the analysis data in \code{x$prepared_data}.
}
\details{
As part of this preparation, the function identifies intervention variables
(\code{A_variables}), time-dependent covariates (\code{B_variables}),
outcome variables (\code{outcome_variables}), competing-risk variables
(\code{competing_variables}), and censoring variables
(\code{censoring_variables}), and stores those names in the object.
}
\examples{
data(simulated_cohort)
ld <- register_format(simulated_cohort)
x <- rtmle_init(time_grid = seq(0,20,4), name_id = "id",
                name_outcome = "stroke", name_competing = "death",
                name_censoring = "dropout",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$timevar_data$stroke[!duplicated(id)],
                   censored_data=ld$timevar_data$dropout,
                   competing_data=ld$timevar_data$death,
                   timevar_data=ld$timevar_data[c("bleeding","changeSBP","A","B")])
x <- add_baseline_data(x,data=ld$baseline_data)
x <- discretize(x,start_followup_date=0)
x <- prepare_rtmle_data(x)
x$prepared_data

}
\seealso{
\code{\link{rtmle_init}}, \code{\link{add_long_data}},
  \code{\link{add_wide_data}}, \code{\link{discretize}},
  \code{\link{protocol}}, \code{\link{target}},
  \code{\link{model_formula}}, \code{\link{run_rtmle}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
