% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rtmle.R
\name{print.rtmle}
\alias{print.rtmle}
\title{Printing an object for register data analysis with the targeted minimum loss estimator}
\usage{
\method{print}{rtmle}(x, ...)
}
\arguments{
\item{x}{Object to be printed}
}
\description{
This print function informs about the current state of the object and
about how to proceed.
}
