% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol<-}
\alias{protocol<-}
\title{Define a named protocol for a hypothetical/emulated trial}
\usage{
protocol(x, ...) <- value
}
\arguments{
\item{x}{An rtmle object as obtained with \code{rtmle_init}.}

\item{...}{Not (yet) used}

\item{value}{list with three forced elements:
\itemize{
\item \code{name}: the name of the protocol
\item \code{treatment_variables}: A vector with the
name(s) of the variable(s) that the protocols intervenes upon. In longitudinal
settings, when a treatment variable is named "A" then the prepared data will contain
a column for each of the variables A_0,A_1,----,A_k. This argument can be left unspecified
in which case the argument \code{intervention} must be a data.frame with names of treatment variables.
An optional argument is
\itemize{
\item \code{verbose}: if FALSE suppress messages
}
See also details.
\item \code{intervention}: A vector, or a named data.frame (tibble, data.table), or a function.
If it is a vector, it should consist of values 0 and 1 corresponding to what the intervention
would set for the \code{treatment_variables}. If it is a data.frame (tibble, data.table),
the names specify the treatment variables and the argument \code{treatment_variables} is ignored.
Each column must be a factor with levels specifying the treatment options. In longitudinal settings,
there can either be only one row in the data.frame and then it is assumed that the intervention is
static throughout the followup period. The data.frame can also have a column named time and
specify for each time point the values that the intervention sets for the treatment variables.
If it is a function, it will be called from \code{intervention_probabilities} with two arguments:
the current time interval and the current history of all variables. The function determines the value(s)
of the treatment variable(s) under the intervention and should return a matrix with as many columns as there are
treatment variables.
}}
}
\value{
The modified object contains the treatment variables and the intervention_table
as list elemens of \code{x$protocols[[name]]} where name is given by \code{value$name}
}
\description{
This function adds a protocol to an existing object.
A protocol defines the values of the treatment variable(s)
at each time point during followup including at time zero (baseline).
}
\examples{
# ------------------------------------------------------------------------------------------
# Intervening on a single treatment variable
# ------------------------------------------------------------------------------------------
x <- rtmle_init(intervals = 3,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
protocol(x) <- list(name = "Always_A",
                    intervention = data.frame("A" = factor("1",levels = c("0","1"))))
protocol(x) <- list(name = "Never_A",
                    intervention = data.frame("A" = factor("0",levels = c("0","1"))))
protocol(x) <- list(name = "Initiate_A_then_stop",
                    intervention = data.frame("A" = factor(c("1","0","0"),levels = c("0","1"))))
x$protocols
# ------------------------------------------------------------------------------------------
# Intervening on a more than one treatment variable
# ------------------------------------------------------------------------------------------
x <- rtmle_init(intervals = 3,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
protocol(x) <- list(name = "Always_A_never_B",
                    intervention = data.frame("A" = factor("1",levels = c("0","1")),
                                              "B" = factor("0",levels = c("0","1"))))
protocol(x) <- list(name = "Always_A_and_B_never_C",
                    intervention = data.frame("A" = factor("1",levels = c("0","1")),
                                              "B" = factor("1",levels = c("0","1")),
                                              "C" = factor("0",levels = c("0","1"))))
x$protocols
}
\author{
Thomas A Gerds \email{tag@biostat.ku.dk}
}
