% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol<-}
\alias{protocol<-}
\title{Define a named protocol for a hypothetical/emulated trial}
\usage{
protocol(x, ...) <- value
}
\arguments{
\item{x}{An rtmle object as obtained with \code{rtmle_init}.}

\item{...}{Not (yet) used}

\item{value}{list with three forced elements:
\itemize{
\item \code{name}: the name of the protocol
\item \code{variable}: the name(s) of the variable(s) that the protocols intervenes upon
\item \code{intervention}: function which determines the value(s) of the variable(s) under the intervention
}}
}
\description{
This function adds a protocol to an existing object.
A protocol defines the values of the treatment variable(s)
at each time point during followup including at time zero (baseline).
}
