% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol}
\alias{protocol}
\title{Define a treatment protocol for an emulated trial}
\usage{
protocol(
  x,
  name,
  intervention,
  expand = TRUE,
  treatment_variables,
  intervene_function = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{rtmle} object as returned by \code{\link{rtmle_init}}.}

\item{name}{Name of the protocol.}

\item{intervention}{A vector, data frame, tibble, or data table specifying
the treatment values dictated by the protocol. If a vector is supplied, it
should contain 0/1 values corresponding to \code{treatment_variables}. If a
data frame, tibble, or data table is supplied, its treatment columns must
be factors whose levels specify the treatment options; in this case
\code{treatment_variables} is ignored. In longitudinal settings, include a
\code{"time"} column in addition to the treatment variables. Ideally,
\code{intervention} supplies values for every value in
\code{x$intervention_nodes}. If fewer rows are supplied and \code{expand =
TRUE}, the last supplied value is used for later time intervals. If
\code{expand = FALSE}, missing time intervals are interpreted as having no
intervention.}

\item{expand}{Logical. If \code{FALSE} and \code{intervention} contains a
\code{"time"} column, do not expand static interventions across time.}

\item{treatment_variables}{A vector with the name(s) of the
variable(s) that the protocol intervenes on. In
longitudinal settings, when a treatment variable is named \code{"A"},
the prepared data contain one column for each of
\code{A_0}, \code{A_1}, \dots, \code{A_k}. This argument can be left
unspecified in which case the argument \code{intervention}
must be given.}

\item{intervene_function}{A character string naming the function used to
intervene under the protocol. Defaults to \code{"intervene"}, which
implements static interventions. The function is called from the internal
function \code{intervention_probabilities()} with two arguments: the
current time interval and the current history of all
variables. It determines the treatment value(s) under the intervention
and should return a matrix with as many columns as there are treatment
variables.}

\item{verbose}{Logical. If \code{FALSE} suppress all messages. \code{TRUE} is the default.}

\item{...}{Not used.}
}
\value{
The modified object contains the treatment variables and
    \code{intervention_table} as list elements of \code{x$protocols[[name]]}.
}
\description{
Adds a named protocol to an existing \code{rtmle} object. A protocol
defines the values of the treatment variable(s) at each time point
during follow-up, including time zero (baseline).
}
\examples{
# ------------------------------------------------------------------------------------------
# Intervening on a single treatment variable
# ------------------------------------------------------------------------------------------
x <- rtmle_init(time_grid=0:3,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- protocol(x,name = "Always_A",
                intervention = data.frame(time=x$intervention_nodes,
                                          "A" = factor("1",levels = c("0","1"))))
x <- protocol(x,name = "Never_A",
              intervention = data.frame(time=x$intervention_nodes,
                                        "A" = factor("0",levels = c("0","1"))))
x <- protocol(x,name = "Initiate_A_then_stop",
              intervention = data.frame(time=x$intervention_nodes,
                                        "A" = factor(c("1","0","0"),levels = c("0","1"))))
x$protocols
# ------------------------------------------------------------------------------------------
# Intervening on more than one treatment variable
# ------------------------------------------------------------------------------------------
x <- rtmle_init(time_grid=0:3,name_id = "id",name_outcome = "Y",name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- protocol(x,name = "Always_A_never_B",
                intervention = data.frame(time=x$intervention_nodes,
                                     "A" = factor("1",levels = c("0","1")),
                                     "B" = factor("0",levels = c("0","1"))))
x <- protocol(x,name = "Always_A_and_B_never_C",
                    intervention = data.frame(time=x$intervention_nodes,
                                              "A" = factor("1",levels = c("0","1")),
                                              "B" = factor("1",levels = c("0","1")),
                                              "C" = factor("0",levels = c("0","1"))))
x$protocols
}
\seealso{
\code{\link{rtmle_init}}, \code{\link{prepare_rtmle_data}},
  \code{\link{intervention_match}}, \code{\link{target}},
  \code{\link{model_formula}}, \code{\link{run_rtmle}}
}
\author{
Thomas A Gerds \email{tag@biostat.ku.dk}
}
