% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmle-package.R
\docType{package}
\name{rtmle-package}
\alias{rtmle}
\alias{rtmle-package}
\title{rtmle: Analysing Longitudinal Observational and Randomized Data with the Targeted Minimum Loss Estimator}
\description{
Flexible implementation of targeted minimum loss estimators. Some functions are optimized for large-scale register data. Some functions in this package were inspired by the 'ltmle' package developed by Joshua Schwab, Samuel Lendle, Maya Petersen, and Mark J. van der Laan, but rtmle provides an independent implementation with a different architecture and extensions to competing risks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tagteam/rtmle}
  \item Report bugs at \url{https://github.com/tagteam/rtmle/issues}
}

}
\author{
\strong{Maintainer}: Thomas Alexander Gerds \email{tag@biostat.ku.dk} (\href{https://orcid.org/0000-0002-5955-816X}{ORCID})

Authors:
\itemize{
  \item Helene Charlotte Wiese Rytgaard Rytgaard \email{hely@sund.ku.dk}
}

Other contributors:
\itemize{
  \item Joshua Schwab (Maintainer and primary author of the ltmle package which inspired this work) [contributor]
  \item Alessandra Meddis \email{alme@sund.ku.dk} [contributor]
  \item Anders Munch \email{a.munch@sund.ku.dk} [contributor]
  \item Kathrine Kold Sørensen \email{kathrine.kold.soerensen@regionh.dk} [contributor]
  \item Christian Torp-Pedersen \email{ctp@heart.dk} [contributor]
  \item Johan Sebastian Ohlendorff \email{johan.sebastian.ohlendorff@sund.ku.dk} [contributor]
}

}
\keyword{internal}
