% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmle-package.R
\docType{package}
\name{rtmle-package}
\alias{rtmle}
\alias{rtmle-package}
\title{rtmle: Analysing Longitudinal Observational and Randomized Data with the Targeted Minimum Loss Estimator}
\description{
Flexible implementation of the targeted minimum loss estimators. Some of the functions are readily optimized for large-scale register data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tagteam/rtmle}
  \item Report bugs at \url{https://github.com/tagteam/rtmle/issues}
}

}
\author{
\strong{Maintainer}: Helene Charlotti Wiese Rytgaard \email{hely@sund.ku.dk}

Authors:
\itemize{
  \item Thomas Alexander Gerds \email{tag@biostat.ku.dk} (\href{https://orcid.org/0000-0002-5955-816X}{ORCID})
}

Other contributors:
\itemize{
  \item Johan Sebastian Ohlendorff \email{johan.sebastian.ohlendorff@sund.ku.dk} [contributor]
  \item Michaela Lukacova \email{michaela.lukacova@sund.ku.dk} [contributor]
}

}
\keyword{internal}
