% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmle-package.R
\docType{package}
\name{rtmle-package}
\alias{rtmle}
\alias{rtmle-package}
\title{rtmle: Analysing Longitudinal Observational and Randomized Data with the Targeted Minimum Loss Estimator}
\description{
Flexible Implementation of the Targeted Minimum Loss Estimators. Functions are Optimized for large-scale Register Data.
}
\author{
\strong{Maintainer}: Thomas Alexander Gerds \email{tag@biostat.ku.dk} (\href{https://orcid.org/0000-0002-5955-816X}{ORCID})

Other contributors:
\itemize{
  \item Johan Sebastian Ohlendorff \email{johan.sebastian.ohlendorff@sund.ku.dk} [contributor]
  \item Michaela Lukacova \email{michaela.lukacova@sund.ku.dk} [contributor]
}

}
\keyword{internal}
