% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmle_init.R
\name{rtmle_init}
\alias{rtmle_init}
\title{Register Targeted Minimum Loss Estimation}
\usage{
rtmle_init(
  time_grid,
  name_id,
  name_time = "date",
  name_outcome,
  name_competing = NULL,
  name_censoring = NULL,
  censored_levels = c("uncensored", "censored"),
  censored_label = "censored",
  minority_threshold = 8,
  weight_truncation = c(0, 1),
  prediction_range = c(0.0001, 0.9999),
  time_grid_labels = time_grid,
  time_unit = NULL
)
}
\arguments{
\item{time_grid}{Vector of time points that define the discrete
time intervals for updating information in a longitudinal
setting. The first value must be 0 and the length of the
vector must be at least 2. When \code{time_grid} is of length
2, no information is updated
after time zero.}

\item{name_id}{Name of the subject identifier variable.}

\item{name_time}{Name of the time variable. For example, \code{"date"} is
appropriate when the data are recorded as calendar dates.}

\item{name_outcome}{Name of the outcome variable or variable prefix. For
example, if \code{name_outcome = "cvddeath"} in a longitudinal setting
with \code{time_grid} of length 4, the wide-format outcome variables are
expected to be named \code{"cvddeath_1"}, \code{"cvddeath_2"}, and
\code{"cvddeath_3"}. If \code{time_grid} has length 2, this should be
the full name of the outcome variable.}

\item{name_competing}{Name of the competing-risk variable or variable
prefix. Can be \code{NULL} if there are no competing risks for the
outcome.}

\item{name_censoring}{Name of the censoring variable or variable prefix.
Can be \code{NULL} if all subjects are uncensored; that is, if the
minimum potential follow-up time is longer than the maximum value of
\code{time_grid}.}

\item{censored_levels}{Character vector with the censoring levels.}

\item{censored_label}{A single character value identifying the censoring
level that indicates censored observations. Must be an element of
\code{censored_levels}.}

\item{minority_threshold}{For binary outcomes, an integer threshold used to
decide whether fitting nuisance-parameter regression models is feasible.
If the number of subjects in the minority group is less than or equal to
this threshold, regression modeling is skipped and the predicted value is
the mean outcome.}

\item{weight_truncation}{Two values defining the minimum and maximum
inverse-probability weights.}

\item{prediction_range}{Two values used to constrain predicted outcome
probabilities to the interval \code{(0, 1)}. This is required for the
logit transformation used in the TMLE fluctuation model. The default is
\code{c(0.0001, 0.9999)}.}

\item{time_grid_labels}{Labels used for the time-grid points on plot
x-axes. Defaults to \code{time_grid}.}

\item{time_unit}{Character string used as the x-axis label in plots. If
\code{NULL} (the default), plots use \code{"Time"}.}
}
\value{
A list with class \code{"rtmle"} and the following elements:
\itemize{
\item targets
\item estimate
\item names
\item times
}
}
\description{
Initialize an rtmle analysis
}
\details{
Creates an empty object of class \code{"rtmle"} that stores analysis
settings and can then be populated with data, regimes, targets, and model
formulas.
}
\examples{
# longitudinal setting
x <- rtmle_init(time_grid=c(0,6,12,18),
                name_id="pnr",
                name_time="date",
                name_outcome="cvddeath",
                name_competing="death",
                name_censoring="Censored",
                censored_levels = c("1","0"),
                censored_label="0")

}
\seealso{
\code{\link{add_baseline_data}}, \code{\link{add_long_data}},
  \code{\link{add_wide_data}}, \code{\link{discretize_data}},
  \code{\link{prepare_rtmle_data}}, \code{\link{regime}},
  \code{\link{target}}, \code{\link{model_formula}},
  \code{\link{run_rtmle}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
