% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rtmle.R
\name{run_rtmle}
\alias{run_rtmle}
\title{Sequential regression with TMLE update step for discretized follow-up data}
\usage{
run_rtmle(
  x,
  targets,
  learner = "learn_glm",
  estimator = "tmle",
  time_horizon,
  refit = TRUE,
  seed = NULL,
  subsets = NULL,
  keep_influence = TRUE,
  progressbar = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{"rtmle"}.}

\item{targets}{Selection of targets to analyze. If missing, all targets in
\code{x$targets} are analyzed.}

\item{learner}{A function, function name, or learner specification used to
fit nuisance-parameter models. Must be one of:
\itemize{
  \item A single string giving the name of a learner function
    (e.g., \code{"learn_glmnet"}) or the function itself.
  \item A list consisting of the elements:
    \itemize{
      \item \code{folds}: a character string giving the name of the
        learner function (e.g., \code{"learn_glmnet"}) or the function itself.
      \item \code{learners}: a named list of learner specifications,
        each of which can either be:
        \itemize{
          \item A single string.
          \item A list containing \code{fun} and parameters to be
            passed to the learner function.
        }
    }
}}

\item{estimator}{Character specifying the estimator: either
\code{"tmle"} or \code{"g-formula"}.}

\item{time_horizon}{The time horizon at which to calculate
risks. If this is a vector, the analysis is performed for
each element.}

\item{refit}{Logical. If \code{TRUE}, ignore any propensity score
and censoring models learned in previous calls to this
function. This may be useful to save computation time. The default
is \code{TRUE}.}

\item{seed}{Seed used for cross-fitting.}

\item{subsets}{A list structure for subset analyses. Each element
is a list requiring a label, to name the subset, and a
subset of the variable \code{x$names$id} in the data
\code{x$prepared_data} to identify the subset. The results of
the subset analysis are stored in
\code{x$estimate[[subsets[[label]]]]}. An optional element of
each subset list is called \code{append}; if \code{TRUE}, append the
estimates to the existing estimates with \code{rbind}. This may be used
for stratified analyses, seed-dependence studies (Monte Carlo error), and
bootstrap analyses. See
examples.}

\item{keep_influence}{Logical. If \code{TRUE}, store the estimated
influence function of the estimator in the object.  Currently
this argument is used only when argument \code{subsets} is also
specified.}

\item{progressbar}{Logical. If \code{TRUE}, show progress of the loops that
fit the nuisance-parameter models.}

\item{verbose}{Logical. If \code{FALSE} suppress all
messages. \code{FALSE} is the default.}

\item{...}{Further arguments can change tuning parameters. For example, use
\code{weight_truncation = c(0.01, 0.99)} to apply weight truncation.}
}
\value{
The modified object contains the fitted nuisance parameter
    models and the estimate of the target parameter.
}
\description{
Runs the analysis defined by \code{\link{rtmle_init}},
\code{\link{protocol}}, \code{\link{target}}, and
\code{\link{model_formula}}.
}
\examples{
# ------------------------------------------------------------------------------------------
# Intervening on a single treatment variable
# ------------------------------------------------------------------------------------------

set.seed(17)
tau <- 3
ld <- simulate_long_data(n = 391,number_visits = 20,
                         beta = list(A_on_Y = -.2,A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = TRUE)
x <- rtmle_init(time_grid = seq(0,1500,30.45*12),name_id = "id",
                name_outcome = "Y",
                name_competing = "Dead",
                name_censoring = "Censored",censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$outcome_data,
                   censored_data=ld$censored_data,
                   competing_data=ld$competing_data,
                   timevar_data=ld$timevar_data)
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x,start_followup_date=0)
x <- protocol(x,name = "Always_A",
                    intervention = data.frame(time=x$intervention_nodes,
                                                   "A" = factor("1",levels = c("0","1"))))
x <- protocol(x,name = "Never_A",
                    intervention = data.frame(time=x$intervention_nodes,
                                              "A" = factor("0",levels = c("0","1"))))
x <- prepare_rtmle_data(x)
x <- target(x,name = "Outcome_risk",
                  estimator = "tmle",
                  protocols = c("Always_A","Never_A"))
x <- model_formula(x)
# default is undersmoothing which means: take the smallest penalty
# where the model still converges
x <- run_rtmle(x,learner = "learn_glmnet",time_horizon = 1:tau)
# with weight truncation
xw <- run_rtmle(x,learner = "learn_glmnet",time_horizon = 1:tau,
weight_truncation=c(0.4,0.6))
# can also use lambda.min or lambda.1se
\dontrun{
    x <- run_rtmle(x,learner = list(name = "glmnet_min",
                                   fun="learn_glmnet",
                                   selector="min"),
                  time_horizon = tau)
summary(x)
}
\dontrun{
# stratified analyses
x <- run_rtmle(x,learner = "learn_glmnet",time_horizon = tau,
               verbose=FALSE,
               subsets=list(list(label="Sex",variable="Sex",
                                 level="Female",id=x$prepared_data[sex==0,id]),
                       list(label="Sex",variable="Sex",
                                 level="Male",id=x$prepared_data[sex==1,id])))
}



}
\seealso{
\code{\link{rtmle_init}}, \code{\link{prepare_rtmle_data}},
  \code{\link{protocol}}, \code{\link{target}}, \code{\link{model_formula}},
  \code{\link{learn_glm}}, \code{\link{learn_glmnet}},
  \code{\link{superlearn}}, \code{\link{summary.rtmle}}
}
\author{
Thomas A Gerds \email{tag@biostat.ku.dk}
}
