% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rtmle.R
\name{run_rtmle}
\alias{run_rtmle}
\title{Sequential regression with TMLE update step for discretized follow-up data}
\usage{
run_rtmle(
  x,
  targets,
  learner = "learn_glm",
  estimator = "tmle",
  time_horizon,
  refit = TRUE,
  seed = NULL,
  subsets = NULL,
  keep_influence = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{targets}{Selection of targets to be analysed. If missing all
targets in x$targets are analysed.}

\item{learner}{A function which is called to fit (learn) the
nuisance parameter models.}

\item{estimator}{Character specifying the estimator: either \code{'tmle'} or \code{'g-formula'}.}

\item{time_horizon}{The time horizon at which to calculate
risks. If it is a vector the analysis will be performed for
each element of the vector.}

\item{refit}{Logical. If \code{TRUE} ignore any propensity score
and censoring models learned in previous calls to this
function. This may be useful to save computation time. Default is \code{TRUE}.}

\item{seed}{Seed used for cross-fitting}

\item{subsets}{A list structure for subset analyses. Each element is a list 
which requires a label, to name the subset, and a subset of the 
variable \code{x$names$id} in the data \code{x$prepared_data} to
identify the subset. The results of the subset analysis are stored
in \code{x$estimate[[subsets[[label]]}. An optional element of each
subset-list is called \code{append}
which should be logical: if \code{TRUE} append the estimates to the existing
estimates with rbind. This may be used for stratified analyses, to study seed dependence (Monte-Carlo error)
and bootstrap. See examples.}

\item{keep_influence}{Logical: if \code{TRUE} store the estimated influence function of the estimator in the object.
Currently this argument is only used when argument \code{subsets} is also specified.}

\item{verbose}{Logical. If \code{FALSE} suppress all messages. \code{FALSE} is the default.}
}
\value{
The modified object contains the fitted nuisance parameter
    models and the estimate of the target parameter.
}
\description{
This function runs the analysis defined in previous steps.
}
\author{
Thomas A Gerds \email{tag@biostat.ku.dk}
}
