% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rtmle.R
\name{run_rtmle}
\alias{run_rtmle}
\title{Sequential regression with TMLE update step for discretized follow-up data}
\usage{
run_rtmle(x, targets, learner = "learn_glm", time_horizon, refit = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{rtmle}}

\item{targets}{Selection of targets to be analysed. If missing all targets in x$targets are analysed.}

\item{learner}{A function which is called to fit (learn) the nuisance parameter models.}

\item{time_horizon}{The time horizon at which to calculate risks. If it is a vector the analysis will be performed for each element of the vector.}

\item{refit}{Logical. If \code{TRUE} ignore any propensity score and censoring models learned in previous calls to this function. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to the learner function.}
}
\description{
This function runs the analysis defined in previous steps.
}
