% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_setting_2.R
\name{sim_data_setting2}
\alias{sim_data_setting2}
\title{Simulate Data in a T2D Diabetes Setting}
\usage{
sim_data_setting2(
  N,
  eta = rep(0.1, 4),
  nu = rep(1.1, 4),
  followup = Inf,
  setting = "a"
)
}
\arguments{
\item{N}{A double of the number of individuals}

\item{eta}{Vector of length 4 of shape parameters for the Weibull intensity with parameterization
\deqn{\eta \nu t^{\nu - 1}}. Default is set to 0.1 for all events.}

\item{nu}{Vector of scale parameters for the Weibull hazard. Default is set to 0.1 for all events.}

\item{followup}{A censoring time. By default set to infinity.}

\item{setting}{Specifies how change in covariate process affects risk of death. Can be set to either "a" or "b".
Is by default set to "a".}
}
\value{
Data frame containing the simulated data. There is a column for ID, time of event (Time),
event type (Delta), baseline covariate (L0), additional covariate (L) and  Treatment Process (A).
}
\description{
Function to simulate data from an T2D diabetes setting. 3 different types of events are simulated, chosen to represent
Death (1), Censoring(2) and Change in Covariate Process(3). Death and Censoring are terminal events and Change in
Covariate Process can occur once.
The intensities of the various events depend upon previous events and the pre specified \eqn{nu} and \eqn{eta}
parameters. The dependence on previous events is controlled by parameters chosen so that a large baseline covariate
(L0) increases the probability of covariate change (L = 1). Initial treatment (A0 = 1) reduces the risk of a change
in the covariate process (L= 1) and the risk of death. A large baseline covariate (L0) increase the risk of death.
probability of operation (A = 1). Operation (A = 1) decreases the risk of death. The censoring process does not
depend on anything. Further two settings can be specified: "a") where the change in the covariate process (L = 1)
increases risk of death, and "b") where the risk of death does not depend on the change in the covariate process (L = 1).
}
\examples{
sim_data_setting2(10)
}
