% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_recur_event_data.R
\name{sim_recur_event_data}
\alias{sim_recur_event_data}
\title{Simulate Survival Data}
\usage{
sim_recur_event_data(N, beta, eta, nu, at_risk, term_deltas)
}
\arguments{
\item{N}{A double for the number of individuals}

\item{beta}{A matrix of doubles for the baseline covariate effects,
the effect of the event number, the treatment effect and the effect of an additional changing covariate L1.}

\item{eta}{Vector of shape parameters}

\item{nu}{Vector of scale parameters}

\item{at_risk}{At risk function}

\item{term_deltas}{Terminal events}
}
\value{
Data frame containing the simulated data
}
\description{
Simulate Survival Data
}
\examples{



N <- 50
eta <- c(0.1, 0.1, 0.1, 0.1)
nu <- c(1.1, 1.1, 1.1, 1.1)
# Effect on Operation
beta0 <- c(3, 0, 1, 9)
# Effect on Censoring
beta1 <- c(0, 0, -1, 5)
# Effect on Death
beta2 <- c(3, 0, 1, 1)
# Effect on time varying covariate
beta3 <- c(3, 0, -1, 0.5)
beta <- cbind(beta0, beta1, beta2, beta3)
term_deltas <- c(1, 2)
at_risk <- function(x, k, m) {
# If you have not died yet or been censored yet, you are at risk for dying or being censored
if(x == 1 | x == 2) return(1)
# You are only at risk for an operation if you have not had an operation yet
else if(x == 0) return(as.numeric(k == 0 | (k == 1 & m == 1)))
# You are only at risk for a change in the covariate process if you have not had a change yet
else return(as.numeric(m == 0))
}
sim_recur_event_data(N = N, beta = beta, eta = eta, nu = nu, at_risk = at_risk, term_deltas = term_deltas)
}
