% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_long_data.R
\name{simulate_long_data}
\alias{simulate_long_data}
\title{library(data.table)
n <- 2000
ld <- simulate_long_data(n,time = d$time,number_epochs=20,beta=0.05)
ld2 <- simulate_long_data(n,time = d$time,number_epochs=20,beta=-0.05)
ld\link{,table(table(id))}
ld\link{,num:=.N,by="id"}
ld\link{num==1}
ld}
\usage{
simulate_long_data(
  n,
  number_epochs,
  baseline_rate,
  beta,
  register_format = FALSE
)
}
\description{
library(data.table)
n <- 2000
ld <- simulate_long_data(n,time = d$time,number_epochs=20,beta=0.05)
ld2 <- simulate_long_data(n,time = d$time,number_epochs=20,beta=-0.05)
ld\link{,table(table(id))}
ld\link{,num:=.N,by="id"}
ld\link{num==1}
ld
}
