% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_long_data.R
\name{simulate_long_data}
\alias{simulate_long_data}
\title{Simulating longitudinal data}
\usage{
simulate_long_data(
  n,
  number_visits = 10,
  baseline_rate,
  beta,
  register_format = FALSE,
  interventional_distribution = FALSE,
  baseline_hazard_outcomes = 0.001
)
}
\arguments{
\item{n}{Sample size}

\item{number_visits}{Number of doctor visit times covariates and treatment change}

\item{baseline_rate}{Vector of hazard rates}

\item{beta}{List of regression coefficients}

\item{register_format}{Logical. If \code{TRUE} the result is not in wide format but re-formatted as a list of register data.}

\item{interventional_distribution}{Logical. If \code{TRUE} the data generating mechanism produces uncensored data under the intervention}

\item{baseline_hazard_outcomes}{Baseline hazard function}
}
\description{
Simulating longitudinal data for illustration purposes
}
\details{
FIXME
}
\examples{
simulate_long_data(10)
}
