% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_long_data.R
\name{simulate_long_data}
\alias{simulate_long_data}
\title{library(data.table)
ld <- simulate_long_data(n=100,number_epochs=20,register_format=TRUE)
map_grid(grid=seq(0,10*timevar_data,
ld\link{,table(table(id))}
ld\link{,num:=.N,by="id"}
ld\link{num==1}
ld}
\usage{
simulate_long_data(
  n,
  number_epochs = 10,
  baseline_rate,
  beta,
  register_format = FALSE,
  int_dist = FALSE,
  baseline_hazard_outcomes = 0.001
)
}
\description{
library(data.table)
ld <- simulate_long_data(n=100,number_epochs=20,register_format=TRUE)
map_grid(grid=seq(0,10*timevar_data,
ld\link{,table(table(id))}
ld\link{,num:=.N,by="id"}
ld\link{num==1}
ld
}
