% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rtmle.R
\name{summary.rtmle}
\alias{summary.rtmle}
\title{Summarize an rtmle analysis}
\usage{
\method{summary}{rtmle}(
  object,
  analysis = "Main_analysis",
  targets,
  reference = NULL,
  digits = 1,
  ...
)
}
\arguments{
\item{object}{Object to summarize.}

\item{analysis}{Name of the analysis to summarize.}

\item{targets}{Names of targets to summarize. Defaults to all targets in the
object.}

\item{reference}{(Optional) Named list of reference protocols, one for each target.}

\item{digits}{Number of decimals for confidence intervals.}

\item{...}{Not used.}
}
\value{
A \code{\link[data.table]{data.table}} with estimates, confidence
  limits, and, when relevant, protocol contrasts.
}
\description{
Summarize rtmle estimates
}
\details{
For each target in the object, summarizes risk estimates and their 95%
confidence limits. If a target includes multiple protocols, the summary also
includes risk differences and risk ratios.
}
\examples{

set.seed(112)
ld <- simulate_long_data(n = 81,number_visits = 20,
                         beta = list(A_on_Y = -.2,
                                     A0_on_Y = -0.3,A0_on_A = 2),
                         register_format = TRUE)

x <- rtmle_init(time_grid = seq(0,1500,30.45*6),name_id = "id",name_outcome = "Y",
                name_competing = "Dead",name_censoring = "Censored",
                censored_label = "censored")
x <- add_long_data(x,
                   outcome_data=ld$outcome_data,
                   censored_data=ld$censored_data,
                   competing_data=ld$competing_data,
                   timevar_data=ld$timevar_data)
x <- add_baseline_data(x,data=ld$baseline_data)
x <- long_to_wide(x)
x <- protocol(x,name = "Always_A",treatment_variables = "A",intervention = 1)
x <- protocol(x,name = "Never_A",treatment_variables = "A",intervention = 0)
x <- prepare_rtmle_data(x)
x <- target(x,name = "Outcome_risk",
                  estimator = "tmle",protocols = c("Always_A","Never_A"))
x <- model_formula(x)
x <- run_rtmle(x,time_horizon=1:2)
summary(x)

}
\seealso{
\code{\link{run_rtmle}}, \code{\link{target}},
  \code{\link{cheap_bootstrap}}, \code{\link{plot.rtmle}}
}
