% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rtmle.R
\name{summary.rtmle}
\alias{summary.rtmle}
\title{Summarizing the results of a register data analysis with the targeted minimum loss estimator}
\usage{
\method{summary}{rtmle}(object, ...)
}
\arguments{
\item{object}{Object to be summarized}
}
\description{
Summarizing the results of a register data analysis with the targeted minimum loss estimator
}
