% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearn.R
\name{superlearn}
\alias{superlearn}
\title{Super learning for rtmle}
\usage{
superlearn(
  folds,
  seed,
  learners,
  parse_learners = TRUE,
  character_formula,
  outcome_variable,
  outcome_variable_name,
  id_variable,
  data,
  intervened_data,
  ensemble_method = "discrete",
  diagnostics,
  ...
)
}
\arguments{
\item{folds}{Number of folds.}

\item{seed}{Random seed.}

\item{learners}{List of learners. Each learner can either be a
character string or a list. If the learner is a character
string it must be the name of a learner function, such as
\code{\link{learn_glm}}, \code{\link{learn_glmnet}},
\code{\link{learn_ranger}}, or \code{\link{learn_xgboost}}. If the
learner is a list, either the name of the learner is a learner function
or the list has an element \code{fun} naming a learner function. The
other elements of the list are passed
as additional arguments to the learner function.}

\item{parse_learners}{Logical. If \code{TRUE}, parse learner
specifications with \code{\link{parse_learners}}.}

\item{character_formula}{A formula (passed as a character string!)
to parse the outcome and the predictor variables.}

\item{outcome_variable}{The name of the outcome variable (same as
\code{all.vars(formula(character_formula))[[1]]}), provided to
avoid overhead in the parsing of the formula.}

\item{outcome_variable_name}{For internal use when called from
\code{\link{run_rtmle}} via \code{fitter()}.}

\item{id_variable}{Name of the subject identifier variable.}

\item{data}{Data used for learning.}

\item{intervened_data}{Data in which all intervention variables have already
been set according to the intervention protocol.}

\item{ensemble_method}{How to combine learners. Implemented methods are
non-linear least squares (\code{"nnls"}), index-of-prediction-accuracy
weighting (\code{"ipa"}), and \code{"discrete"}, which selects the learner
with the lowest Brier score. If all models have zero weight, the average
predicted value from the learning set is used.}

\item{diagnostics}{For internal use when called from \code{\link{run_rtmle}}
via \code{fitter()}, which is called from
\code{intervention_probabilities()} and \code{sequential_regression()}.}

\item{...}{Not used.}
}
\value{
The level-one data column of the discrete super learner.
}
\description{
Super learn nuisance-parameter models
}
\details{
Uses cross-fitting to produce level-one data for learners used by
\code{\link{run_rtmle}}. The predicted probability from each learner is
stored as one column and compared with the outcome. The discrete super
learner chooses the learner with the lowest quadratic prediction error; the
ensemble super learner combines the candidate learners.
}
\examples{
# Note that the function is designed to be called from run_rtmle.
library(ranger)
library(glmnet)
data(simulated_cohort)
ld <- register_format(simulated_cohort)
A0 <- ld$timevar_data$A[date == 0, list(id, A_0 = value)]
d <- merge(ld$baseline_data, A0, by = "id")
fit <- superlearn(folds = 2,seed = 8,
                  learners = list("learn_glm",
                                  "my_ranger" = list(fun = "learn_ranger",
                                                     num.trees = 5),
                                  "learn_glmnet" = list(alpha = 1,
                                                     fun="learn_glmnet")),
                  character_formula = "A_0~sex+age+SBP",
                  outcome_variable = "A_0",
                  id_variable = "id",
                  data = d,
                  intervened_data = d)

}
\seealso{
\code{\link{run_rtmle}}, \code{\link{parse_learners}},
  \code{\link{learn_glm}}, \code{\link{learn_glmnet}},
  \code{\link{learn_ranger}}, \code{\link{learn_xgboost}}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
