% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearn.R
\name{superlearn}
\alias{superlearn}
\title{Super learning for rtmle}
\usage{
superlearn(
  folds,
  seed,
  learners,
  character_formula,
  outcome_variable,
  id_variable,
  data,
  intervened_data,
  ensemble,
  ...
)
}
\arguments{
\item{folds}{Number of folds}

\item{seed}{Random seed}

\item{learners}{List of learners. Each learner can either be a character string or a list. If the learner is a
character string it must be the name of a learner function, such as \code{\link{learn_glm}},
\code{\link{learn_glmnet}},  \code{\link{learn_ranger}}. If the learner is a list then either the name of the learner
is a learner function or the list has an element \code{learner_fun} which is the name of a learner function. The other elements of the list are
passed on as additional arguments to the learner function.}

\item{character_formula}{A formula (passed as a character string!) to parse the outcome and the predictor variables.}

\item{outcome_variable}{The name of the outcome variable (same as \code{all.vars(formula(character_formula))[[1]]} provided to avoid overhead in the parsing of the formula.}

\item{id_variable}{The name of the subject id variable.}

\item{data}{The data for learning.}

\item{intervened_data}{The data were all intervention variables are readily set according to the intervention protocol.}

\item{ensemble}{Type of superlearning. When \code{"winner"} the discrete super learner is the learner with the lowest
level-one mean squared prediction error (Brier score).}

\item{...}{Not (yet) used.}
}
\value{
The level-one data column of the discrete superlearner.
}
\description{
Super learning nuisance parameter models for longitudinal TMLE analyses
}
\details{
The function collaborates with \code{run_rtmle}. It
It uses cross-fitting to produce level-one data, where the predicted probability of each learner
is a column which can be compared with the outcome. The discrete super learner chooses the
learner lowest quadratic prediction error. The ensemble super learner combines the learners.
}
\examples{
# Note that the function is designed to be called from run_rtmle.  
library(ranger)
library(glmnet)
set.seed(17)
ld <- simulate_long_data(n = 113,
                         number_visits = 20,
                         beta = list(A_on_Y = -.2,
                                     A0_on_Y = -0.3,A0_on_A = 6),
                         register_format = FALSE)
fit <- superlearn(folds = 2,seed = 8,
                  learners = list("learn_glm",
                                  "my_ranger" = list(learner_fun = "learn_ranger",
                                                     num.trees = 5),
                                  "learn_glmnet" = list(alpha = 1)),
                  character_formula = "A_0~L_0+sex+age",
                  outcome_variable = "A_0",
                  id_variable = "id",
                  data = ld[!duplicated(id)],
                  intervened_data = ld[!duplicated(id)])

}
\seealso{
\code{\link{run_rtmle}}, \code{learn_glm}, \code{learn_glmnet}, \code{learn_ranger}
}
\author{
Thomas A. Gerds \href{mailto:tag@biostat.ku.dk}{tag@biostat.ku.dk}
}
