% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Define a target parameter and specify formulas for the estimators of the nuisance parameters}
\usage{
target(x, name, protocols, estimator, ...)
}
\arguments{
\item{x}{An rtmle object as obtained with \code{rtmle_init}.}

\item{name}{the name of the target parameter.}

\item{protocols}{The names of the protocol(s) involved in the target parameters.}

\item{estimator}{Character specifying the estimator: either \code{'tmle'} or \code{'g-formula'}.}

\item{...}{Not (yet) used}
}
\description{
A target defines a target parameter, an estimator and models
for the nuisance parameters.
}
\examples{
x <- rtmle_init(time_grid = 0:2, name_id = "id", name_outcome = "Y")
x <- protocol(x, name = "Always_A",
              intervention = data.frame(time = x$intervention_nodes,
                                        A = factor("1", levels = c("0", "1"))))
x <- protocol(x, name = "Never_A",
              intervention = data.frame(time = x$intervention_nodes,
                                        A = factor("0", levels = c("0", "1"))))
x <- target(x, name = "Outcome_risk",
            protocols = c("Always_A", "Never_A"),
            estimator = "tmle")
x$targets
x$estimator
}
