% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Define a target parameter and specify formulas for the estimators of the nuisance parameters}
\usage{
target(x, name, protocols, estimator, ...)
}
\arguments{
\item{x}{An rtmle object as obtained with \code{rtmle_init}.}

\item{name}{the name of the target parameter.}

\item{protocols}{The names of the protocol(s) involved in the target parameters.}

\item{estimator}{Character specifying the estimator: either \code{'tmle'} or \code{'g-formula'}.}

\item{...}{Not (yet) used}
}
\description{
A target defines a target parameter, an estimator and models
for the nuisance parameters.
}
